### 2021-08-12 #3604 @danielskatz
👋 @rouson & @tclune - Would one of you be able to review this submission for JOSS? I'm particularly looking for one reviewer with expertise in Fortran good/best practices around packaging, testing, distribution, etc. Or do you have suggestions for someone else in this area?
