### 2021-08-16 #3604 @michaeltouati
Dear Tom,

I can recommend this list of potential reviewer :

- @sigvaldm ***@***.*** ***@***.***>)
- @CFGrote ***@***.*** ***@***.***>)
- @joglekara ***@***.*** ***@***.***>)
- @TomGoffrey ***@***.*** ***@***.***>)

Thanking you,
Best regards,
Michaël

> On 16 Aug 2021, at 15:01, Tom Clune ***@***.***> wrote:
>
>
> @danielskatz <https://github.com/danielskatz> I was hoping @rouson <https://github.com/rouson> would blink first. :-) I'm a bit swamped at work over the next couple of weeks - if the review can be delayed until early/mid September, you can consider me signed on.
>
