### 2021-08-16 #3604 @danielskatz
👋 @sigvaldm, @CFGrote, @joglekara, @TomGoffrey - Would one of you be willing to review this submission to JOSS?
