### 2021-08-16 #3618 @TomGoffrey
The code appears to compile and run correctly. However, when running `make extract` I get several pages of warnings:

```
MatplotlibDeprecationWarning: Passing parameters norm and vmin/vmax simultaneously is deprecated since 3.3
and will become an error two minor releases later. Please pass vmin/vmax directly to the norm when creating
it.
```

and

```
MatplotlibDeprecationWarning: shading='flat' when X and Y have the same dimensions as C is deprecated
since 3.3.  Either specify the corners of the quadrilaterals with X and Y, or pass shading='auto',
'nearest' or 'gouraud', or set rcParams['pcolor.shading'].  This will become an error two minor
releases later.
```

These are fairly trivial to fix, and would improve the user experience. They also won't be warnings forever.

I also have a query about the author list (@danielskatz you might want to comment on this?) - The JOSS guidelines state

```
Author lists must be correct and complete. All listed authors must have made a contribution to the work,
and all significant contributors should be included in the author list.
```

It looks like a second user made the initial commit, or at least at some point made a single very large commit. Could you confirm that the current (single) author list conforms to the above guideline?

A note on the paper- the Beam Warming method should be capitalised (after R. F. Warming and R. M. Beam).

I'll continue my review, but wanted to bring these to your attention in the meantime.
