### 2021-08-17 #3618 @TomGoffrey
There's also a floating point exception when compiled in debug mode, running with the input deck as shipped in the main directory:

```
Program received signal SIGFPE: Floating-point exception - erroneous arithmetic operation.

Backtrace for this error:
#0  0x7ff518c6b2ed in ???
#1  0x7ff518c6a503 in ???
#2  0x7ff5182d103f in ???
#3  0x7ff518938ea4 in ???
#4  0x7ff5188c11fe in ???
#5  0x55b21ea2f0e5 in __library_MOD_init_simu
	at sources/library.f90:115
#6  0x55b21ea31e78 in esvm
	at sources/esvm.f90:47
#7  0x55b21ea32c8a in main
	at sources/esvm.f90:11
```
