### 2021-08-18 #3618 @danielskatz
Re the author question from @TomGoffrey

> I also have a query about the author list (@danielskatz you might want to comment on this?) - The JOSS guidelines state

>>Author lists must be correct and complete. All listed authors must have made a contribution to the work,
and all significant contributors should be included in the author list.

>It looks like a second user made the initial commit, or at least at some point made a single very large commit. Could you confirm that the current (single) author list conforms to the above guideline?

I'm curious to hear what @michaeltouati says before I comment.
