### 2021-08-22 #3618 @michaeltouati
Toi whom it may concern,

	I am the only one contributor to this version of the code.

My friend, Nitin Shukla, added the openACC instruction for parallelization on GPUs on another version. I've made the mistake to accept its commit without checking it was working for all test cases and I’ve noticed it wasn’t. I thus came back to the original version after that. You can ask him.

It would be a pleasure to add him in the list of authors but he didn’t maker any contribution to this version of the code.

Best regards,
Michaël

> On 18 Aug 2021, at 15:03, Daniel S. Katz ***@***.***> wrote:
>
>
> Re the author question from @TomGoffrey <https://github.com/TomGoffrey>
> I also have a query about the author list ***@***.*** <https://github.com/danielskatz> you might want to comment on this?) - The JOSS guidelines state
>
> Author lists must be correct and complete. All listed authors must have made a contribution to the work,
> and all significant contributors should be included in the author list.
>
> It looks like a second user made the initial commit, or at least at some point made a single very large commit. Could you confirm that the current (single) author list conforms to the above guideline?
>
> I'm curious to hear what @michaeltouati <https://github.com/michaeltouati> says before I comment.
>
