### 2021-08-30 michaeltouati/ESVM#5 @tclune
No automated tests
This issue is part of the openjournal review.

Ideally there should be a step like `make tests` that checks that correct values are returned for some test case that covers a reasonable fraction of the software.

Referencing: openjournals/joss-reviews#3618
