### 2021-09-08 #3618 @danielskatz
👋 @tclune, @TomGoffrey, @rouson - we're now about 3 weeks into the process. How is it going? I see some progress by @tclune and @TomGoffrey, but no checked items from @rouson
