### 2021-09-12 #3618 @danielskatz
@tclune - can you mention `openjournals/joss-reviews#3618` in your issue so that it (and its status) is visible here?
