### 2021-09-13 #3618 @TomGoffrey
@tclune Whilst it's true that the 3rd issue doesn't directly block any of the review checklists, I'd personally rather see it fixed during the review process. Those warnings will become errors in future versions of matplotlib. Similar for the FPE after compiling in debug mode. Doesn't directly block any of the checklist, but clearly not desirable.
