### 2021-09-13 michaeltouati/ESVM#9 @TomGoffrey
Improvements to JOSS Paper
Part of openjournals/joss-reviews#3618

I'd like to suggest a few minor improvements to the paper:

 - [x] Captilise "Beam-Warming", inline with other eponymous schemes.
 - [x] Add references to similar codes in the "Statement of Need" section. This section currently reads as a description of the code, and future development plans, whereas for the JOSS review requires a statement of need which makes clear what the need for the code is, as well as a "State of the field".
 - [x] I'm not familiar with the use of "linear" and "non-linear" to describe the two lists of solvers they use. From my perspective they appear to be schemes without and with TVD limiters. Perhaps this could be clarified.
 - [x] Move the list of proposed additions to the code to a later point in the paper.
 - [x] Expand on the description of each of the Figures. Currently there's only a list. For example, are these run with the input decks as provided (if so, include a git hash in case they change), how can the user verify these results are correct?
 - [x] In the discussion of velocity bins, perhaps an example and explanation as used in the production of one of the figures might be useful to a novice user.
