### 2021-09-22 #3618 @michaeltouati
Dear Daniel,

	Sorry for my late reply. I was on vacations and I start today a two day presencial formation.
I’ll be back at my lab. on Friday and make all the needed corrections / improvements by the end of the week.

I’ve indeed seen and followed that Tom Clune rose some issues that, when I will fixe them, will make the code and the paper much better.

Thanking you,
Best regards,
Michaël

> On 20 Sep 2021, at 15:34, Daniel S. Katz ***@***.***> wrote:
>
>
> 👋 @michaeltouati <https://github.com/michaeltouati> - any news on work on the open issues?
>
