+++ sources/library.f90
@@ -90,14 +90,14 @@ pure subroutine INIT_SIMU(x, vx, f_n)
     dx  = 0.25_PR  ! "particle size"
     dvx = 0.025_PR ! "particle size"
     xs  = (x_max - x_min) / 8.  
-	  do l=1,N_vx,1
-  		do i=1,N_x,1
+    do l=1,N_vx,1
+      do i=1,N_x,1
         f_n(i,l) = (1.0_PR/sqrt(2._PR*pi*(vs**2._PR)))*&
         & exp(-(vx(l)**2._PR)/(2._PR*(vs**2._PR))) 
         X2 = -0.5_PR * ( ( ( (x(i)-xs) / dx )**2._PR) + ( ( (vx(l)-vd) / dvx )**2._PR) )
         f_n(i,l) = f_n(i,l) + ( A * exp(X2)/ (2._PR * pi * dx * dvx ) )
-	  	end do
-  	end do
+      end do
+    end do
   ! Landau damping test case :
   else if (perturb == 2) then
     do l=1,N_vx,1
@@ -127,21 +127,21 @@ pure subroutine INIT_SIMU(x, vx, f_n)
   ! Boundary conditions
   select case (b_cond)
     ! absorbing
-  	case (1)
-	  	do l=-1,N_vx+2,1
+    case (1)
+      do l=-1,N_vx+2,1
         f_n(N_x+1,l) = 0._PR
         f_n(N_x+2,l) = 0._PR
         f_n(0,l)     = 0._PR
         f_n(-1,l)    = 0._PR
-  		end do
+      end do
     ! periodic
-  	case (2)
-	 		do l=-1,N_vx+2,1
+    case (2)
+      do l=-1,N_vx+2,1
         f_n(N_x+1,l) = f_n(1,l)
         f_n(N_x+2,l) = f_n(2,l)
         f_n(0,l)     = f_n(N_x,l)
         f_n(-1,l)    = f_n(N_x-1,l)
-  		end do
+      end do
   end select
   do i=-1,N_x+2,1
     f_n(i,N_vx+1) = 0._PR
@@ -187,10 +187,10 @@ subroutine DENSITIES(vx, f_n, n_e, j_e, v_e, vT_e)
   do i=-1,N_x+2,1
     allocate(F1(1:N_vx),F2(1:N_vx),F3(1:N_vx))
     F1(1:N_vx) = f_n(i,1:N_vx)*d_vx
-  	do l=1,N_vx,1
+    do l=1,N_vx,1
       F2(l) = - f_n(i,l) * vx(l) * d_vx
       F3(l) = f_n(i,l) * (vx(l)**2._PR) * d_vx
-  	end do
+    end do
     n_e(i)  = sum(F1(1:N_vx))
     j_e(i)  = sum(F2(1:N_vx))
     v_e(i)  = - j_e(i) / n_e(i)
@@ -233,10 +233,10 @@ subroutine POISSON(N_t, d_t, d_x, x, j_e, n_e, E_x_n, E_x_np1, phi_n)
     b(i)        =  2._PR
     c(i)        = -1._PR
     d(i)        = (1._PR - n_e(i))*(d_x**2._PR)
-  	if (b_cond.eq.2) then
+    if (b_cond.eq.2) then
       e(i)        = 0._PR
       phi_temp(i) = 0._PR
-  	end if
+    end if
   end do
   !omp END PARALLEL DO
   select case (b_cond)
@@ -358,7 +358,7 @@ subroutine DRIVE(N_t, d_t, time, d_x, x, E_x_n, E_x_np1, phi_n)
 end subroutine DRIVE
 
[…]
