+++ sources/input.f90
@@ -52,189 +52,191 @@ real(PR), public  :: Te
 contains
 
 subroutine read_init_parameters
-implicit none
-Character(len=80)         :: str
-Integer                   :: i, istr
-integer                   :: N_th_max
+  implicit none
+  Character(len=80)         :: str
+  Integer                   :: i, istr
+  integer                   :: N_th_max
 
-! read the file 'input-deck'
-open (unit=1, file='input-deck',&
-& position='rewind', access='sequential',&
-& form='formatted', action='read',status='old')
-read: do
-call get_str(str)
-if (str == 'end') exit read
-if (str(1:1) /= '#') cycle read
-istr = 80
-do i = 1, len(str)
-if (str(i:i) == ' ') then
-istr = i
-exit
-end if
-end do
-select case (str(1:istr))
+  ! read the file 'input-deck'
+  open (unit=1, file='input-deck',&
+  & position='rewind', access='sequential',&
+  & form='formatted', action='read',status='old')
 
-case ('#N_th')
-N_th = get_integer(str(istr+1:))
-case ('#T')  
-T = get_real(str(istr+1:))
-case ('#Z')  
-Z = get_real(str(istr+1:))
-case ('#ni')  
-ni = get_real(str(istr+1:))
+  read: do
+    
+    call get_str(str)
+    
+    if (str == 'end') exit read
+    
+    if (str(1:1) /= '#') cycle read
+    
+    istr = 80
+    do i = 1, len(str)
+      if (str(i:i) == ' ') then
+        istr = i
+        exit
+      end if
+    end do
+    
+    select case (str(1:istr))
+      case ('#N_th')
+        N_th = get_integer(str(istr+1:))
+      case ('#T')  
+        T = get_real(str(istr+1:))
+      case ('#Z')  
+        Z = get_real(str(istr+1:))
+      case ('#ni')  
+        ni = get_real(str(istr+1:))
+      case ('#x_min')  
+        x_min = get_real(str(istr+1:))
+      case ('#x_max')  
+        x_max = get_real(str(istr+1:))
+      case ('#d_x')  
+        d_x = get_real(str(istr+1:))
+      case ('#vx_min')  
+        vx_min = get_real(str(istr+1:))
+      case ('#vx_max')  
+        vx_max = get_real(str(istr+1:))
+      case ('#d_vx')  
+        d_vx = get_real(str(istr+1:))
+      case ('#cfl')
+        cfl = get_real(str(istr+1:))  
+      case ('#L_t')
+        L_t = get_real(str(istr+1:))
+      case ('#dt_diag')
+        dt_diag = get_real(str(istr+1:))
+      case ('#maxwell')
+        maxwell = get_integer(str(istr+1:))
+      case ('#b_cond')
+        b_cond = get_integer(str(istr+1:))
+      case ('#scheme')
+        scheme = get_integer(str(istr+1:))
+      case ('#b')
+        b      = get_real(str(istr+1:))
+      case ('#perturb')
+        perturb = get_integer(str(istr+1:))
+      case ('#A')
+        A = get_real(str(istr+1:))
+      case ('#k')
+        k = get_real(str(istr+1:))
+      case ('#omega_0')
+        omega_0= get_real(str(istr+1:))
+      case ('#vd')
+        vd = get_real(str(istr+1:))
+      case ('#vs')
+        vs = get_real(str(istr+1:))
+    end select
 
-case ('#x_min')  
-x_min = get_real(str(istr+1:))
-case ('#x_max')  
-x_max = get_real(str(istr+1:))
[…]
