+++ sources/extract.py
@@ -167,10 +167,17 @@ for i in range(0,Nx):
 cmap = plt.get_cmap(cmap)
 Maxval = np.amax(p)
 Minval = np.amin(p)
+if ( Minval == Maxval ) :
+    if ( Minval == 0. ) :
+        Maxval =  1.
+        Minval = -1.
+    else :
+        Maxval = 1.1*Maxval
+        Minval = 0.9*Maxval
 norm = cm.colors.Normalize(vmax=Maxval, vmin=Minval)
 fig=plt.figure()
 plt.rc('text', usetex=True)
-plt.pcolormesh(X,T,P,cmap=cmap,norm=norm,vmax=Maxval,vmin=Minval)
+plt.pcolormesh(X,T,P,cmap=cmap,norm=norm,shading='auto')
 cbar=plt.colorbar()
 cbar.ax.tick_params(labelsize=16)
 plt.title(title, fontdict=font)
@@ -217,10 +224,17 @@ for i in range(0,Nx):
 cmap = plt.get_cmap(cmap)
 Maxval = np.amax(p)
 Minval = np.amin(p)
+if ( Minval == Maxval ) :
+    if ( Minval == 0. ) :
+        Maxval =  1.
+        Minval = -1.
+    else :
+        Maxval = 1.1*Maxval
+        Minval = 0.9*Maxval
 norm = cm.colors.Normalize(vmax=Maxval, vmin=Minval)
 fig=plt.figure()
 plt.rc('text', usetex=True)
-plt.pcolormesh(X,T,P,cmap=cmap,norm=norm,vmax=Maxval,vmin=Minval)
+plt.pcolormesh(X,T,P,cmap=cmap,norm=norm,shading='auto')
 cbar=plt.colorbar()
 cbar.ax.tick_params(labelsize=16)
 plt.title(title, fontdict=font)
@@ -285,10 +299,17 @@ for line in file:
         cmap = plt.get_cmap(cmap)
         Maxval = np.amax(p[(int(N)-1)*int(N3):int(N)*int(N3)])
         Minval = np.amin(p[(int(N)-1)*int(N3):int(N)*int(N3)])
+        if ( Minval == Maxval ) :
+            if ( Minval == 0. ) :
+                Maxval =  1.
+                Minval = -1.
+            else :
+                Maxval = 1.1*Maxval
+                Minval = 0.9*Maxval
         norm = cm.colors.Normalize(vmax=Maxval, vmin=Minval)
         fig=plt.figure()
         plt.rc('text', usetex=True)
-        plt.pcolormesh(X,VX,P,cmap=cmap,norm=norm,vmax=Maxval,vmin=Minval)
+        plt.pcolormesh(X,VX,P,cmap=cmap,norm=norm,shading='auto')
         cbar=plt.colorbar()
         cbar.ax.tick_params(labelsize=16)
         plt.title(title+' at '+str(time)+r'$\,\omega_p^{-1}$', fontdict=font)
@@ -342,7 +363,16 @@ for line in file:
         plt.xlim([np.amin(X),np.amax(X)])
         plt.ylabel(r'$E_x\,(m_e \omega_p v_T / e)$', fontdict=font)
         plt.yticks(fontsize=16)
-        plt.ylim([np.amin(p),np.amax(p)])
+        ex_min = np.amin(p)
+        ex_max = np.amax(p)
+        if ( ex_min == ex_max ):
+            if (ex_min == 0.) :
+                ex_min = -1.
+                ex_max =  1.
+            else :
+                ex_min = 0.9 * ex_max
+                ex_max = 1.1 * ex_max
+        plt.ylim([ex_min,ex_max])
[…]
