### 2021-09-27 ee288e0
First improvements of the paper
- Beam-Warming capitalized
- TVD property added
- Perspectives moved to the end of the paper
