+++ paper.md
@@ -9,7 +9,7 @@ tags:
   - Vlasov equation
   - donor cell
   - Lax-Wendroff
-  - beam warming
+  - Beam-Warming
   - Fromm 
   - minmod
   - superbee
@@ -38,16 +38,12 @@ bibliography: paper.bib
 
 # Summary
 
-ESVM (ElectroStatic Vlasov-Maxwell) is a 1D-1V Vlasov-Maxwell Fortran code parallelized using OpenMP and developed 
-to adapt simulations to specific Plasma Physics problems with explicit linear finite volume numerical advection schemes such as the donor cell, the Lax-Wendroff, the beam warming or the Fromm method and non-linear ones such as the minmod, the superbee, the Van Leer, the MUSCL1 or the MUSCL2 method coupled to the Maxwell-Gauss equation for the electrostatic field or the Maxwell-Ampere equation with Poisson equation computed at the first time step. Python scripts, using matplotlib and numpy packages, are provided to automatically extract the simulation results, to plot them and to save them. Compilation rules can be easily modified depending on the user compiler preferences using the provided makefile.
+ESVM (ElectroStatic Vlasov-Maxwell) is a Vlasov-Maxwell Fortran 90 standard compliant code, parallelized using OpenMP and developed 
+to adapt simulations to specific Plasma Physics problems with explicit linear finite volume numerical advection schemes such as the donor cell, the Lax-Wendroff, the Beam-Warming or the Fromm method and non-linear ones such as the minmod, the superbee, the Van Leer, the MUSCL1 or the MUSCL2 method coupled to the Maxwell-Gauss equation for the electrostatic field or the Maxwell-Ampere equation with Poisson equation computed at the first time step. Python scripts, using matplotlib and numpy packages, are provided to automatically extract the simulation results, to plot them and to save them. Compilation rules can be easily modified depending on the user compiler preferences using the provided makefile. 
 
 # Statement of need
 
-`ESVM` allows for adapting simulations to a specific Plasma Physics problem : it is possible to chose a linear finite volume numerical advection schemes such as the donor cell, the Lax-Wendroff, the beam warming or the Fromm method or a non-linear one such as the minmod, the superbee, the Van Leer, the MUSCL1 or the MUSCL2 method in order to compute the Vlasov equation phase-space derivatives and to chose between computing the Poisson equation versus computing the Maxwell-Ampere equation (with Poisson equation computed at the first time step only). Well known academic Plasma Physics problems have been used to validate the code and are provided as examples (input deck and obtained simulation results). It is planned to :
-- extend the code to relativistic 2D-2V phase-space simulations
-- implement the Perfectly Matched Layer technique for Electromagnetic Fields absorption at the spatial simulation box boundaries
-- implement its MPI parallelization
-- implement the Belyaev-Budker relativistic collision operator
+`ESVM` allows for adapting simulations to a specific Plasma Physics problem : it is possible to chose a linear finite volume numerical advection schemes such as the donor cell, the Lax-Wendroff, the Beam-Warming or the Fromm method or a non-linear schemes using total variation diminishing (TVD) flux limiters such as the minmod, the superbee, the Van Leer, the MUSCL1 or the MUSCL2 method in order to compute the Vlasov equation phase-space derivatives and to chose between computing the Poisson equation versus computing the Maxwell-Ampere equation (with Poisson equation computed at the first time step only). Well known academic Plasma Physics problems have been used to validate the code and are provided as examples (input deck and obtained simulation results). The code can be used by Plasma Physics undergradurate students, teachers and researchers.
 
[…]
