### 2021-09-28 4bcc27b
Fix Compilation debug warnings
