### 2021-09-28 fd250c9
Fix Compilation debug warnings
