+++ makefile
@@ -10,7 +10,7 @@
 # openMP #
 ##########
 
-#OPTS = -prec-div -prec-sqrt -openmp -openmp-report2 -r8
+#OPTS = -openmp -r8 -i8 -O3
 
 #########
 # debug #
@@ -30,7 +30,7 @@ F90 = gfortran
 # openMP #
 ##########
 
-# OPTS = -fopenmp -O3 -ffixed-line-length-none -fdefault-integer-8 -fdefault-real-8
+OPTS = -fopenmp -fdefault-real-8 -fdefault-integer-8 -O3
 
 #########
 # debug #
@@ -50,7 +50,11 @@ SRC_PATH_PY = ${SRC_PATH}
 
 SRCS        = acuracy.f90 constants.f90 physics.f90 input.f90 diagnostics.f90 library.f90 esvm.f90
 
+SRCS_CHK    = acuracy.f90 constants.f90 physics.f90 input.f90 input-chk.f90
+
 OBJTS := $(SRCS:%.f90=%.o)
+
+OBJTS_CHK := acuracy.o constants.o physics.o input.o input-chk.o
 
 all : ESVM
 
@@ -71,6 +75,9 @@ physics.o : $(SRC_PATH)physics.f90
 input.o : $(SRC_PATH)input.f90
 	$(F90) $(OPTS) -c $(SRC_PATH)input.f90
 
+input-chk.o : $(SRC_PATH)input-chk.f90
+	$(F90) $(OPTS) -c $(SRC_PATH)input-chk.f90
+
 diagnostics.o : $(SRC_PATH)diagnostics.f90
 	$(F90) $(OPTS) -c $(SRC_PATH)diagnostics.f90
 
@@ -79,6 +86,26 @@ library.o : $(SRC_PATH)library.f90
 	
 esvm.o : $(SRC_PATH)esvm.f90
 	$(F90) $(OPTS) -c $(SRC_PATH)esvm.f90
+
+########################
+########################
+##  CHECK INPUT DECK  ##
+########################
+########################
+
+check-input-deck : $(OBJTS_CHK)
+	$(F90) $(OPTS) $(OBJTS_CHK) -o check-input-deck
+	rm *.mod
+	rm *.o
+
+check : check-input-deck
+	./check-input-deck
+
+################
+################
+##  CLEANING  ##
+################
+################
 
 clean :
 	rm -rf *.o *.mod esvm
@@ -97,3 +124,121 @@ extract : $(SRC_PATH_PY)extract.py
 
 extract_logfe : $(SRC_PATH_PY)extract_f_log.py
 	python3 $(SRC_PATH_PY)extract_f_log.py
+
+#############
+#############
+##  TESTS  ##
+#############
+#############
+
+RED=$(shell tput setaf 1)
+GREEN=$(shell tput setaf 2)
+RESET=$(shell tput sgr0)
+
+test_ampere :
+	@echo '--------------------------------'
+	@echo '            ${RED}WARNING${RESET}          '
+	@echo '--------------------------------'
+	@echo 'The code should be compiled with'
+	@echo 'double floating point precisions'
+	@echo '     for the tests to pass!     '
+	@echo '--------------------------------'
+	@echo '         Maxwell solver         '
+	@echo '--------------------------------'
+	@echo -n 'Maxwell-Ampere solver  : '
+	@cp test-cases/Test/Ampere/input-deck .
+	@./esvm > test.output
+	@diff test.output test-cases/Test/Ampere/output; \
+    TST=$$?;\
+    if [ $$TST -eq 0 ]; then echo -n "${GREEN}PASSED${RESET}"; else echo -n "${RED}NOT PASSED${RESET}"; fi; echo ' '; \
+
+test_poisson :
[…]
