+++ sources/input-chk.f90
@@ -1,0 +1,17 @@
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+!!                                                                   !!
+!!          1D-1V ElectroStatic Vlasov-Maxwell (ESV) code            !!
+!!                                                                   !!
+!!  Written by Dr Michaël J TOUATI - CLPU - 2020 - mtouati@clpu.es   !!
+!!                                                                   !!
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+!
+program CHECK_INPUT_DECK
+!
+use acuracy
+use constants
+use input
+!
+call read_init_parameters()
+!
+end program CHECK_INPUT_DECK
