+++ makefile
@@ -163,8 +163,9 @@ test_ampere :
 	@echo '   the corresponding simulation '
 	@echo '   performed by the developper  '
 	@echo '   located in test-cases/Test/  '
-	@echo ' concerning a plasma at Maxwell-'
-	@echo ' Boltzmann equilibrium          ' 
+	@echo ' concerning a simple plasma     '
+	@echo ' simulation at Maxwell-Boltzmann'
+	@echo ' equilibrium                    ' 
 	@echo '--------------------------------'
 	@echo '         Maxwell solver         '
 	@echo '--------------------------------'
