+++ sources/diagnostics.f90
@@ -60,21 +60,21 @@ subroutine DIAG_ENERGY(time, N_x, d_x, n_e, v_e, vT_e, E_x_n, &
   end do
   !
   if (abs(U_K).lt.zero) then
-    write (8,'(4E23.15)') time, zero
+    write (8,'(4E22.14)') time, zero
   else
-    write(8,'(4E23.15)') time, U_K
+    write(8,'(4E22.14)') time, U_K
   end if
 
   if (abs(U_T).lt.zero) then
-    write (9,'(4E23.15)') time, zero
+    write (9,'(4E22.14)') time, zero
   else
-    write(9,'(4E23.15)') time, U_T
+    write(9,'(4E22.14)') time, U_T
   end if
 
   if (abs(U_E).lt.zero) then
-    write (10,'(4E23.15)') time, zero
+    write (10,'(4E22.14)') time, zero
   else
-    write(10,'(4E23.15)') time, U_E
+    write(10,'(4E22.14)') time, U_E
   end if
 end subroutine DIAG_ENERGY
 
@@ -92,69 +92,69 @@ subroutine DIAG(N_t, time, N_x, x, N_vx, vx, test_positivity, U_K, U_T, U_E, &
   real(PR)                               , intent(in) :: U_K, U_T, U_E
   integer                                             :: l, i
   !
-  print*,'==========================='
-  print*,'time (/omega_p) =', time 
-  print*,'Number of iteration :',N_t
-  print*,'==========================='
-  print*, ' '
+  write (*,*)'==========================='
+  write (*,'(A,1E22.14)')'time (/omega_p) =', time 
+  write (*,'(A,I7)')'Number of iteration :',N_t
+  write (*,*)'==========================='
+  write (*,*) ' '
   if (test_positivity.eqv..true.) print*, 'the distribution function became negative'
-  print*,'Kinetic energy  (n0 Debye^3 me v_T^2 / 2) = ', U_K
-  print*,'Thermal energy  (n0 Debye^3 me v_T^2 / 2) = ', U_T
-  print*,'Electric energy (n0 Debye^3 me v_T^2 / 2) = ', U_E
-  print*,'--------------------------------------------'
-  print*,'Total energy    (n0 Debye^3 me v_T^2 / 2) = ', U_K + U_T + U_E 
-  print*, ' '
+  write (*,'(A,1E22.14)')'Kinetic energy  (n0 Debye^3 me v_T^2 / 2) = ', U_K
+  write (*,'(A,1E22.14)')'Thermal energy  (n0 Debye^3 me v_T^2 / 2) = ', U_T
+  write (*,'(A,1E22.14)')'Electric energy (n0 Debye^3 me v_T^2 / 2) = ', U_E
+  write (*,*)'--------------------------------------------'
+  write (*,'(A,1E22.14)')'Total energy    (n0 Debye^3 me v_T^2 / 2) = ', U_K + U_T + U_E 
+  write (*,*) ' '
   !
   do l=1,N_vx,1
     do i = 1,N_x,1
       if (abs(f_n(i,l)).lt.zero) then
-        write (1,'(4E23.15)') time, vx(l), x(i), zero
+        write (1,'(4E22.14)') time, vx(l), x(i), zero
       else
-        write(1,'(4E23.15)') time, vx(l), x(i), f_n(i,l)
+        write(1,'(4E22.14)') time, vx(l), x(i), f_n(i,l)
       end if
     end do
   end do
   do i = 1,N_x,1
     if (abs(n_e(i)).lt.zero) then
-      write (2,'(4E23.15)') time, x(i), zero
+      write (2,'(4E22.14)') time, x(i), zero
     else
[…]
