+++ makefile
@@ -144,17 +144,11 @@ RESET=$(shell tput sgr0)
 
 test_ampere :
 	@echo '--------------------------------'
-	@echo '            ${RED}WARNING${RESET}          '
+	@echo '        TESTS DESCRIPTION       '
 	@echo '--------------------------------'
-	@echo 'The code should be compiled with'
+	@echo '${RED}The code should be compiled with'
 	@echo 'double floating point precisions'
-	@echo '     for the tests to pass!     '
-	@echo '--------------------------------'
-	@echo '            ${RED}WARNING${RESET}          '
-	@echo '--------------------------------'
-	@echo 'The code should be compiled with'
-	@echo 'double floating point precisions'
-	@echo '     for the tests to pass!     '
+	@echo '     for the tests to pass!${RESET}     '
 	@echo 'The tests consist in performing '
 	@echo ' diff file1 file2 where :       '
 	@echo ' * file1 is the test simulation '
@@ -273,9 +267,9 @@ test_MUSCL2 :
     TST=$$?;\
     if [ $$TST -eq 0 ]; then echo "${GREEN}PASSED${RESET}"; else echo "${RED}NOT PASSED${RESET}"; fi; echo ' '; \
     rm -f test.output
-	rm -rf results/
+	@rm -rf results/
 
-test : test_ampere test_poisson \
+test : test_ampere test_poisson test_openMP \
 	   test_donor_cell test_Lax_Wendroff test_Beam_Warming test_Fromm \
-	   test_minmod test_superbee test_Van_Leer test_MUSCL1 test_MUSCL2
+ 	   test_minmod test_superbee test_Van_Leer test_MUSCL1 test_MUSCL2
 
