+++ sources/input.f90
@@ -159,49 +159,51 @@ subroutine read_init_parameters
     stop
   end if
 
-  print*,'-----------------------------------------'
-  print*,'Recapitulation of simulation parameters :'
-  print*,'-----------------------------------------'
-  print*,'* Number of OpenMP threads :'
-  print*,'N_th    = ',N_th
-  print*,'-----------------------------------------'
-  print*,'* Plasma properties (immobile ions)'
-  print*,'T       = ',T
-  print*,'Z       = ',Z
-  print*,'ni      = ',ni
-  print*,'-----------------------------------------'
-  print*,'* 1D-1V phase-space : '
-  print*,'x_min   = ',x_min
-  print*,'x_max   = ',x_max
-  print*,'d_x     = ',d_x
-  print*,'vx_min  = ',vx_min
-  print*,'vx_max  = ',vx_max
-  print*,'d_vx    = ',d_vx
-  print*,'-----------------------------------------'
-  print*,'* Simulation properties : '
-  print*,'cfl     = ',cfl
-  print*,'L_t     = ',L_t
-  print*,'dt_diag = ',dt_diag
-  print*,'maxwell = ',maxwell
-  print*,'b_cond  = ',b_cond
-  print*,'scheme  = ',scheme
-  print*,'b       = ',b
-  print*,'-----------------------------------------'
-  print*,'* Perturbation properties : '
-  print*,'perturb = ',perturb
-  print*,'A       = ',A
-  print*,'k       = ',k
-  print*,'omega_0 = ',omega_0
-  print*,'vd      = ',vd
-  print*,'vs      = ',vs
-  print*,'-----------------------------------------'
-  print*,'* Deduced parameters : '
-  print*,'N_x     = ',N_x
-  print*,'N_vx    = ',N_vx
-  print*,'n0      = ',n0
-  print*,'Te      = ',Te
-  print*,'-----------------------------------------' 
-  print*, ' '
+  write (*,*)'-----------------------------------------'
+  write (*,*)'Recapitulation of simulation parameters :'
+  write (*,*)'-----------------------------------------'
+  write (*,*)'* Number of OpenMP threads :'
+  write (*,'(A,1I4)')'N_th    = ',N_th
+  write (*,*)'-----------------------------------------'
+  write (*,*)'* Plasma properties (immobile ions)'
+  write (*,'(A,1E22.14)')'T       = ',T
+  write (*,'(A,1E22.14)')'Z       = ',Z
+  write (*,'(A,1E22.14)')'ni      = ',ni
+  write (*,*)'-----------------------------------------'
+  write (*,*)'* 1D-1V phase-space : '
+  write (*,'(A,1E22.14)')'x_min   = ',x_min
+  write (*,'(A,1E22.14)')'x_max   = ',x_max
+  write (*,'(A,1E22.14)')'d_x     = ',d_x
+  write (*,'(A,1E22.14)')'vx_min  = ',vx_min
+  write (*,'(A,1E22.14)')'vx_max  = ',vx_max
+  write (*,'(A,1E22.14)')'d_vx    = ',d_vx
+  write (*,*)'-----------------------------------------'
+  write (*,*)'* Simulation properties : '
+  write (*,'(A,1E22.14)')'cfl     = ',cfl
+  write (*,'(A,1E22.14)')'L_t     = ',L_t
+  write (*,'(A,1E22.14)')'dt_diag = ',dt_diag
+  write (*,'(A,1I1)')'maxwell = ',maxwell
+  write (*,'(A,1I1)')'b_cond  = ',b_cond
+  write (*,'(A,1I2)')'scheme  = ',scheme
[…]
