+++ sources/diagnostics.f90
@@ -60,21 +60,21 @@ subroutine DIAG_ENERGY(time, N_x, d_x, n_e, v_e, vT_e, E_x_n, &
   end do
   !
   if (abs(U_K).lt.zero) then
-    write (8,'(4E22.14)') time, 0.
+    write(8,'(4E23.15)') time, zero
   else
-    write(8,'(4E22.14)') time, U_K
+    write(8,'(4E23.15)') time, U_K
   end if
-
+  !
   if (abs(U_T).lt.zero) then
-    write (9,'(4E22.14)') time, 0.
+    write(9,'(4E23.15)') time, zero
   else
-    write(9,'(4E22.14)') time, U_T
+    write(9,'(4E23.15)') time, U_T
   end if
-
+  !
   if (abs(U_E).lt.zero) then
-    write (10,'(4E22.14)') time, 0.
+    write(10,'(4E23.15)') time, zero
   else
-    write(10,'(4E22.14)') time, U_E
+    write(10,'(4E23.15)') time, U_E
   end if
 end subroutine DIAG_ENERGY
 
@@ -94,85 +94,75 @@ subroutine DIAG(N_t, time, N_x, x, N_vx, vx, test_positivity, U_K, U_T, U_E, &
   integer                                             :: l, i
   !
   write (*,*)'==========================='
-  write (*,'(A,1E22.14)')'time (/omega_p) =', time 
+  write (*,'(A,1E11.3)')'time (/omega_p) =',time 
   write (*,'(A,1I7)')'Number of iteration :',N_t
   write (*,*)'==========================='
-  write (*,*) ' '
+  write (*,*)' '
   if (test_positivity.eqv..true.) print*, 'the distribution function became negative'
-  if (abs(U_K).lt.zero) then
-    write (*,'(A,1E22.2)')'Kinetic energy  (n0 Debye^3 me v_T^2 / 2) = ', U_K
-  else
-    write (*,'(A,1E22.2)')'Kinetic energy  (n0 Debye^3 me v_T^2 / 2) = ', 0.
-  end if
-  if (abs(U_T).lt.zero) then
-    write (*,'(A,1E22.2)')'Thermal energy  (n0 Debye^3 me v_T^2 / 2) = ', U_T
-  else
-    write (*,'(A,1E22.2)')'Thermal energy  (n0 Debye^3 me v_T^2 / 2) = ', 0.
-  end if
-  if (abs(U_E).lt.zero) then
-    write (*,'(A,1E22.2)')'Electric energy (n0 Debye^3 me v_T^2 / 2) = ', U_E
-  else
-    write (*,'(A,1E22.2)')'Electric energy (n0 Debye^3 me v_T^2 / 2) = ', 0.
-  end if
-  write (*,*)'--------------------------------------------'
+  write (*,'(A,1E8.1)')'Kinetic energy  (n0 Debye^3 me v_T^2 / 2) = ', U_K
+  write (*,'(A,1E8.1)')'Thermal energy  (n0 Debye^3 me v_T^2 / 2) = ', U_T
+  write (*,'(A,1E8.1)')'Electric energy (n0 Debye^3 me v_T^2 / 2) = ', U_E
+  write (*,*)'---------------------------------------------------'
   U_tot = U_K + U_T + U_E 
-  if (abs(U_tot).lt.zero) then
-    write (*,'(A,1E22.2)')'Total energy    (n0 Debye^3 me v_T^2 / 2) = ', U_tot
-  else
-    write (*,'(A,1E22.2)')'Total energy    (n0 Debye^3 me v_T^2 / 2) = ', 0.
-  end if
-  write (*,*) ' '
[…]
