+++ makefile
@@ -4,19 +4,19 @@
 #############
 #############
 
-F90 = ifort
+# F90 = ifort
 
 ##########
 # openMP #
 ##########
 
-OPTS = -fopenmp -prec-div -prec-sqrt -mp1 -r8 -O3 
+# OPTS = -fopenmp -r8 -O3
 
 #########
 # debug #
 #########
 
-#OPTS = -g -fpe0 -ftrapuv -debug all -debug-parameters all -C -traceback -openmp -openmp-report2 -r8
+# OPTS = -g -traceback -fopenmp -r8 -std90 -fpe0 -debug all -debug-parameters all -C 
 
 ################
 ################
@@ -24,19 +24,19 @@ OPTS = -fopenmp -prec-div -prec-sqrt -mp1 -r8 -O3
 ################
 ################
 
-#F90 = gfortran
+F90 = gfortran
 
 ##########
 # openMP #
 ##########
 
-#OPTS = -fopenmp -fdefault-real-8 -O3
+OPTS = -fopenmp -fdefault-real-8 -O3
 
 #########
 # debug #
 #########
 
-#OPTS = -fdefault-real-8 -O -g -fopenmp -Wall -fcheck=all -fbacktrace -ffpe-trap=invalid,zero,overflow
+# OPTS = -fdefault-real-8 -O -g -fopenmp -Wall -fcheck=all -fbacktrace -std=f95 -fall-intrinsics -ffpe-trap=invalid,zero,overflow
 
 #####################################
 #####################################
@@ -147,9 +147,6 @@ test_ampere :
 	@echo '        TESTS DESCRIPTION       '
 	@echo '--------------------------------'
 	@echo '                                '
-	@echo '${RED}The code should be compiled with'
-	@echo 'double floating point precisions'
-	@echo '     for the tests to pass!${RESET}     '
 	@echo 'The tests consist in performing '
 	@echo ' diff file1 file2 where :       '
 	@echo ' * file1 is the test simulation '
@@ -193,6 +190,26 @@ test_openMP :
     TST=$$?;\
     if [ $$TST -eq 0 ]; then echo "${GREEN}PASSED${RESET}"; else echo "${RED}NOT PASSED${RESET}"; fi; echo ' '; \
 
+test_periodic :
+	@echo '--------------------------------'
+	@echo '      Boundary conditions       '
+	@echo '--------------------------------'
+	@echo '                                '
+	@echo -n 'Periodic bound. cond.  : '
+	@cp test-cases/Test/Boundary-conditions/Periodic/input-deck .
+	@./esvm > test.output
+	@diff test.output test-cases/Test/Boundary-conditions/Periodic/output; \
+    TST=$$?;\
+    if [ $$TST -eq 0 ]; then echo "${GREEN}PASSED${RESET}"; else echo "${RED}NOT PASSED${RESET}"; fi; echo ' '; \
+
+test_absorbing :
+	@echo -n 'Absorbing bound. cond. : '
+	@cp test-cases/Test/Boundary-conditions/Absorbing/input-deck .
+	@./esvm > test.output
+	@diff test.output test-cases/Test/Boundary-conditions/Absorbing/output; \
+    TST=$$?;\
+    if [ $$TST -eq 0 ]; then echo "${GREEN}PASSED${RESET}"; else echo "${RED}NOT PASSED${RESET}"; fi; echo ' '; \
+
 test_donor_cell :
 	@echo '--------------------------------'
[…]
