+++ sources/physics.f90
@@ -17,21 +17,21 @@ public :: omega_pe, v_T, Debye
   
 contains
 
-elemental function omega_pe(n0)
+function omega_pe(n0)
   implicit none
   real(PR), intent(in) :: n0       ! (/cm^3)
   real(PR)             :: omega_pe ! (/s)
   omega_pe = sqrt(4 * pi * n0 * (e**2._PR) / me)
 end function omega_pe
   
-elemental function v_T(Te)
+function v_T(Te)
   implicit none
   real(PR), intent(in) :: Te  ! Te (K)
   real(PR)             :: v_T ! v_T (cm/s)
   v_T = (3._PR* kb * Te / me )**0.5_PR
 end function v_T
   
-elemental function Debye(n0, Te)
+function Debye(n0, Te)
   implicit none
   real(PR), intent(in) :: n0, Te ! ne (/cm^3) and T (K)
   real(PR)             :: Debye  ! Debye (cm)
