### 2021-09-29 #3618 @michaeltouati
@danielskatz - If you have such a job for me, I'm in. :).  My current contract indeed ends this December and I'm actively searching for a new job.

@tclune  and @TomGoffrey  - Thank you very much for your review. As you suggested :
- I've made the code Fortran 90 standard compliant
- I've solved the different Matplotlib Deprecation Warning issues
- I've capitalized "Beam-Warming", inline with other eponymous schemes (this is like they chose the numerical scheme they developed in agreement with their family name!)
- I've mentioned the non-linear solvers use a TVD flux limiter
- I moved the perspectives of code additions later in the paper
- I have added in the community guidelines for who this code can be useful (under and post-graduate students, teachers and researchers)
- I've solved the compiling warning error when compiled in debug mode (it was caused by double precision underflows)
- I've added a make test in the makefile (that copy and run pre-prepared input-decks and compared with the corresponding simulation by comparing the terminal output containing the different energy contributions and total energy in each simulation using a diff). I've tested it with gnu and intel compilers and all tests pass if they are compiled with double precision.
- I've added a make check that allows the user to check the simulation parameters he has filled in the input deck

I'll improve the paper.md tomorrow and I'll try to finish before you go on vacation @tclune . Following your suggestions :
- I'll add add references in the paper.bib of similar codes and Plasma Physics needed literature and papers related with the  academic cases that I provide and that are located in test-cases/
- I'll expand the description of figures by adding the derivation of analytical estimates for the different academic cases that I provide and I'll explain that, by running the same input-deck that I provide, the reader/user should obtain the same figures that I provide. Indeed, I didn't add the simulation data output of the academic cases in order to limit the size of the code to ~100 MB and not GBs (it is a kinetic code!). Also, I'll add a discussion about the velocity bin sizes there.
- Finally, I'll add in the readme and in the paper.md a description of the make check and make test that I have added.

Again, thank you very much for your suggestions / corrections. I think the code and the paper will be greatly improved thanks to you. I have been happy to learn it is already currently used by a PhD student who found the code thanks to GitHub and use it for his PhD investigations in Theoretical Plasma Physics.

Best regards,
Michaël
