### 2021-09-29 michaeltouati/ESVM#5 @michaeltouati
I have added automated tests.
The user just need to type make test after having compiled the code.
I have tested it : all tests pass if the code is compiled with double precision
