### 2021-09-30 #3618 @rouson
@danielskatz I was just getting started and noticed that all of the checkboxes on my list appear to be grayed out.  I also see that I'm not listed under "Assigned," whereas the other two reviewers are listed so I assume you have to assign this issue to me for me to review it.
