+++ paper.bib
@@ -1,1 +1,15 @@
+@book{GaleevSagdeev:1969,
+  author = {Sagdeev, R. Z. and Galeev, A. A.},
+  title = "{Nonlinear Plasma Theory}",
+  year = "1969",
+  publisher = {W. A. Benjamin, Inc., New York}
+}
 
+@book{Landau1981,
+  Author = {Landau, L. D. and Lifshitz, E. M.},
+  Series = {{A Course of Theoretical Physics}},
+  Title = {{Physical kinetics}},
+  Volume = {10},
+  Year = {1981}
+  publisher = {Pergamon Press}
+}
