+++ paper.bib
@@ -1,3 +1,16 @@
+@article{VanLeer:1977,
+  title = {Towards the ultimate conservative difference scheme III. Upstream-centered finite-difference schemes for ideal compressible flow},
+  journal = {Journal of Computational Physics},
+  volume = {23},
+  number = {3},
+  pages = {263-275},
+  year = {1977},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1016/0021-9991(77)90094-8},
+  url = {https://www.sciencedirect.com/science/article/pii/0021999177900948},
+  author = {Van Leer, B.}
+}
+
 @book{GaleevSagdeev:1969,
   author = {Sagdeev, R. Z. and Galeev, A. A.},
   title = "{Nonlinear Plasma Theory}",
