+++ paper.bib
@@ -1,3 +1,18 @@
+@article{LaxWendroff:1960,
+  author = {Lax, Peter and Wendroff, Burton},
+  title = {Systems of conservation laws},
+  journal = {Communications on Pure and Applied Mathematics},
+  volume = {13},
+  number = {2},
+  pages = {217-237},
+  doi = {https://doi.org/10.1002/cpa.3160130205},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cpa.3160130205},
+  eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/cpa.3160130205},
+  year = {1960}
+}
+
+
+
 @article{VanLeer:1977,
   title = {Towards the ultimate conservative difference scheme III. Upstream-centered finite-difference schemes for ideal compressible flow},
   journal = {Journal of Computational Physics},
