+++ paper.bib
@@ -1,3 +1,30 @@
+@article{Courant:1952,
+  author = {Courant, R and Isaacson, E and Rees, M},
+  title = {On the solution of nonlinear hyperbolic differential equations by finite differences},
+  journal = {Communications on Pure and Applied Mathematics},
+  volume = {5},
+  number = {3},
+  pages = {243-255},
+  doi = {https://doi.org/10.1002/cpa.3160050303},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cpa.3160050303},
+  eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/cpa.3160050303},
+  year = {1952}
+}
+
+@Article{Godunov:1959,
+  Author = {Godunov, S. K.},
+  Title = {{Eine Differenzenmethode f\"ur die N\"aherungsberechnung unstetiger L\"osungen der hydrodynamischen Gleichungen}},
+  FJournal = {{Matematicheski\u{\i} Sbornik. Novaya Seriya}},
+  Journal = {{Mat. Sb., Nov. Ser.}},
+  Volume = {47},
+  Pages = {271--306},
+  Year = {1959},
+  Publisher = {Izdatel'stvo Nauka, Moskva},
+  Language = {Russian},
+  MSC2010 = {76M20 76L05},
+  Zbl = {0171.46204}
+}
+
 @article{LaxWendroff:1960,
   author = {Lax, Peter and Wendroff, Burton},
   title = {Systems of conservation laws},
@@ -11,7 +38,31 @@
   year = {1960}
 }
 
+@article{Fromm:1968,
+  title = {A method for reducing dispersion in convective difference schemes},
+  journal = {Journal of Computational Physics},
+  volume = {3},
+  number = {2},
+  pages = {176-189},
+  year = {1968},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1016/0021-9991(68)90015-6},
+  url = {https://www.sciencedirect.com/science/article/pii/0021999168900156},
+  author = {Fromm, J. E.}
+}
 
+@article{BeamWarming:1976,
+  title = {An implicit finite-difference algorithm for hyperbolic systems in conservation-law form},
+  journal = {Journal of Computational Physics},
+  volume = {22},
+  number = {1},
+  pages = {87-110},
+  year = {1976},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1016/0021-9991(76)90110-8},
+  url = {https://www.sciencedirect.com/science/article/pii/0021999176901108},
+  author = {Beam, R. M. and Warming, R.F }
+}
 
 @article{VanLeer:1977,
   title = {Towards the ultimate conservative difference scheme III. Upstream-centered finite-difference schemes for ideal compressible flow},
@@ -24,6 +75,19 @@
   doi = {https://doi.org/10.1016/0021-9991(77)90094-8},
   url = {https://www.sciencedirect.com/science/article/pii/0021999177900948},
   author = {Van Leer, B.}
+}
+
+@article{Roe:1986,
+  author = {Roe, P L},
+  title = {Characteristic-Based Schemes for the Euler Equations},
+  journal = {Annual Review of Fluid Mechanics},
+  volume = {18},
+  number = {1},
+  pages = {337-365},
+  year = {1986},
+  doi = {10.1146/annurev.fl.18.010186.002005},
+  URL = {https://doi.org/10.1146/annurev.fl.18.010186.002005},
+  eprint = {https://doi.org/10.1146/annurev.fl.18.010186.002005}
 }
 
 @book{GaleevSagdeev:1969,
