+++ paper.bib
@@ -11,11 +11,11 @@
   year = {1952}
 }
 
-@Article{Godunov:1959,
+@article{Godunov:1959,
   Author = {Godunov, S. K.},
-  Title = {{Eine Differenzenmethode f\"ur die N\"aherungsberechnung unstetiger L\"osungen der hydrodynamischen Gleichungen}},
+  Title = {Eine Differenzenmethode f\"ur die N\"aherungsberechnung unstetiger L\"osungen der hydrodynamischen Gleichungen},
   FJournal = {{Matematicheski\u{\i} Sbornik. Novaya Seriya}},
-  Journal = {{Mat. Sb., Nov. Ser.}},
+  Journal = {Mat. Sb., Nov. Ser.},
   Volume = {47},
   Pages = {271--306},
   Year = {1959},
@@ -99,9 +99,9 @@
 
 @book{Landau1981,
   Author = {Landau, L. D. and Lifshitz, E. M.},
-  Series = {{A Course of Theoretical Physics}},
+  Series = {A Course of Theoretical Physics},
   Title = {{Physical kinetics}},
   Volume = {10},
-  Year = {1981}
+  Year = {1981},
   publisher = {Pergamon Press}
 }
