+++ paper.bib
@@ -38,6 +38,19 @@
   year = {1960}
 }
 
+@article{Yee:1966,
+  Author = {Kane Yee},
+  Doi = {10.1109/TAP.1966.1138693},
+  Issn = {0018-926X},
+  Journal = {IEEE Transactions on Antennas and Propagation},
+  Keywords = {EMP radiation effects.;Magnetic scattering by absorbing media;Boundary value problems;Maxwell equations;Difference equations;Finite difference methods;Boundary conditions;Electromagnetic scattering;Partial differential equations;EMP radiation effects;Conductors;Differential equations},
+  Month = {May},
+  Number = {3},
+  Pages = {302-307},
+  Volume = {14},
+  Year = {1966},
+  Bdsk-Url-1 = {https://doi.org/10.1109/TAP.1966.1138693}}
+
 @article{Fromm:1968,
   title = {A method for reducing dispersion in convective difference schemes},
   journal = {Journal of Computational Physics},
@@ -97,7 +110,7 @@
   publisher = {W. A. Benjamin, Inc., New York}
 }
 
-@book{Landau1981,
+@book{Landau:1981,
   Author = {Landau, L. D. and Lifshitz, E. M.},
   Series = {A Course of Theoretical Physics},
   Title = {{Physical kinetics}},
