+++ paper.bib
@@ -38,6 +38,19 @@
   year = {1960}
 }
 
+@article{Dawson:1962,
+  author = {Dawson,John },
+  title = {One‐Dimensional Plasma Model},
+  journal = {The Physics of Fluids},
+  volume = {5},
+  number = {4},
+  pages = {445-459},
+  year = {1962},
+  doi = {10.1063/1.1706638},
+  URL = {https://aip.scitation.org/doi/abs/10.1063/1.1706638},
+  eprint = {https://aip.scitation.org/doi/pdf/10.1063/1.1706638}
+}
+
 @article{Yee:1966,
   Author = {Kane Yee},
   Doi = {10.1109/TAP.1966.1138693},
@@ -102,6 +115,48 @@
   URL = {https://doi.org/10.1146/annurev.fl.18.010186.002005},
   eprint = {https://doi.org/10.1146/annurev.fl.18.010186.002005}
 }
+
+@article{Tzoufras:2011,
+  title = {A Vlasov–Fokker–Planck code for high energy density physics},
+  journal = {Journal of Computational Physics},
+  volume = {230},
+  number = {17},
+  pages = {6475-6494},
+  year = {2011},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1016/j.jcp.2011.04.034},
+  url = {https://www.sciencedirect.com/science/article/pii/S0021999111002828},
+  author = {M. Tzoufras and A.R. Bell and P.A. Norreys and F.S. Tsung},
+  keywords = {Numerical methods, High energy density physics, Vlasov, Fokker–Planck, Laser fusion, Electron transport, Spherical harmonics}
+}
+
+@article{Touati:2014,
+	doi = {10.1088/1367-2630/16/7/073014},
+	url = {https://doi.org/10.1088/1367-2630/16/7/073014},
+	year = 2014,
+	month = {jul},
+	publisher = {{IOP} Publishing},
+	volume = {16},
+	number = {7},
+	pages = {073014},
+	author = {M Touati and J-L Feugeas and Ph Nicolaï and J J Santos and L Gremillet and V T Tikhonchuk},
+	title = {A reduced model for relativistic electron beam transport in solids and dense plasmas},
+	journal = {New Journal of Physics}
+}
+
+@article{Joglekar:2020,
+  doi = {10.21105/joss.02182},
+  url = {https://doi.org/10.21105/joss.02182},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {53},
+  pages = {2182},
+  author = {Archis S. Joglekar and Matthew C. Levy},
+  title = {VlaPy: A Python package for Eulerian Vlasov-Poisson-Fokker-Planck Simulations},
+  journal = {Journal of Open Source Software}
+}
+
 
 @book{GaleevSagdeev:1969,
   author = {Sagdeev, R. Z. and Galeev, A. A.},
