+++ paper.md
@@ -38,23 +38,21 @@ bibliography: paper.bib
 
 # Summary
 
-ESVM (ElectroStatic Vlasov-Maxwell) is a Vlasov-Maxwell Fortran 90 code, parallelized using OpenMP and developed in order to adapt simulations to specific Plasma Physics problems thanks to different finite volume numerical advection schemes @Godunov:1959 implemented in the code :
-- donor-cell scheme i.e. the downwind scheme or upwind one @Courant:1952 depending on the advection direction in each cell, 
+ESVM (ElectroStatic Vlasov-Maxwell) is a Vlasov-Maxwell Fortran 90 code and parallelized using OpenMP allowing for the study of collisionless plasmas . The different finite volume numerical advection schemes @Godunov:1959 implemented in the code in order to discretize the Vlasov equation are the :
+- donor-cell scheme i.e. the downwind / upwind scheme @Courant:1952 depending on the advection direction in each phase-space cell, 
 - Lax-Wendroff scheme @LaxWendroff:1960, 
 - Fromm scheme @Fromm:1968,
 - Beam-Warming scheme @BeamWarming:1976,
+- Van Leer scheme @VanLeer:1977
 - minmod scheme @Roe:1986, 
-- superbee scheme @Roe:1986, 
-- Van Leer scheme @VanLeer:1977 and 
-- two resulting Monotonic Upwind-centered Scheme for Conservation Laws (MUSCL) schemes MUSCL1 and MUSCL2. 
+- superbee scheme @Roe:1986 and 
+- two Monotonic Upwind-centered Scheme for Conservation Laws (MUSCL) schemes MUSCL1 and MUSCL2. 
 
-make use of total variation diminishing (TVD) flux limiters. The discretized Vlasov equation is coupled with the self-consistent Maxwell-Gauss equation for the electrostatic field or equivalently with the Maxwell-Ampere equation with Poisson equation computed at the first time step. While the discretized Poisson equation needs the inversion of a triangular matrix, the second one make use of a faster second order finite difference numerical scheme. Python scripts, using matplotlib and numpy packages, are provided to automatically extract and plot the simulation results that are stored in text files. Compilation rules can be easily modified depending on the user compiler preferences using the provided makefile. 
+Contrary to the linear second order Lax-Wendroff, Fromm and Beam-Warming schemes, the non-linear second order minmod, superbee, Van Leer and MUSCL schemes make use of a Total Variation Diminishing (TVD) flux limiter with the price of becoming a first order scheme in some cell with large quantity oscillation. The discretized Vlasov equation is coupled with the self-consistent Maxwell-Gauss equation for the electrostatic field or equivalently with the Maxwell-Ampere equation with Poisson equation computed at the first time step. While the 1D discretized Poisson equation needs the inversion of a triangular matrix, the second one makes use of the faster second order finite difference Yee scheme @Yee:1966. Python scripts, using the Matplotlib and Numpy packages, are provided to automatically extract and plot the simulation results that are stored in text files. Compilation rules can be easily modified depending on the user compiler preferences using the provided makefile. 
 
 # Statement of need
 
[…]
