+++ paper.md
@@ -48,11 +48,11 @@ ESVM (ElectroStatic Vlasov-Maxwell) is a Vlasov-Maxwell Fortran 90 code and para
 - superbee scheme @Roe:1986 and 
 - two Monotonic Upwind-centered Scheme for Conservation Laws (MUSCL) schemes MUSCL1 and MUSCL2. 
 
-Contrary to the linear second order Lax-Wendroff, Fromm and Beam-Warming schemes, the non-linear second order minmod, superbee, Van Leer and MUSCL schemes make use of a Total Variation Diminishing (TVD) flux limiter with the price of becoming a first order scheme in some cell with large quantity oscillation. The discretized Vlasov equation is coupled with the self-consistent Maxwell-Gauss equation for the electrostatic field or equivalently with the Maxwell-Ampere equation with Poisson equation computed at the first time step. While the 1D discretized Poisson equation needs the inversion of a triangular matrix, the second one makes use of the faster second order finite difference Yee scheme @Yee:1966. Python scripts, using the Matplotlib and Numpy packages, are provided to automatically extract and plot the simulation results that are stored in text files. Compilation rules can be easily modified depending on the user compiler preferences using the provided makefile. 
+Contrary to the linear second order Lax-Wendroff, Fromm and Beam-Warming schemes, the non-linear second order minmod, superbee, Van Leer and MUSCL schemes make use of a Total Variation Diminishing (TVD) flux limiter with the price of becoming a first order scheme in some phase-space cells to limit numerical oscillations. The donor-cell scheme is a first order scheme. It has the pros of naturally limiting such eventual oscillations but the cons of being highly numerically diffusive so. In ESVM, the discretized Vlasov equation is coupled with the self-consistent Maxwell-Gauss equation for the electrostatic field or equivalently with the Maxwell-Ampere equation with Poisson equation computed at the first time step. While the 1D discretized Poisson equation needs the inversion of a triangular matrix, the second one makes use of the faster second order finite difference Yee scheme @Yee:1966. Python scripts, using the Matplotlib and Numpy packages, are provided to automatically extract and plot the simulation results that are stored in text files. Compilation rules can be easily modified depending on the user compiler preferences using the provided makefile. 
 
 # Statement of need
 
-`ESVM` has been developed in order to adapt simulations to specific Plasma Physics problems by chosing the more adequate finite volume numerical advection schemes to compute the Vlasov equation phase-space derivatives and to chose between computing the Poisson equation versus computing the Maxwell-Ampere equation (with Poisson equation computed at the first time step only). Well known academic Plasma Physics problems have been used to validate the code and are provided as examples (input deck and obtained simulation results). The code aims at beeing used by the Plasma Physics open-source Highly Parallel Computing (HPC) community ranging from under and post-graduate students to teachers and theoretical researchers who usually use Particle-In-Cell (PIC) codes @Dawson:1962 to study collisionless plasmas. However, the PIC method, that can be seen as a Lagrangian Vlasov equation solver, may prohibit the study of Physical processes on large time scales and/or for dense collisionless plasmas due to the fluctuations of computed quantities imposed by the use of a finite number of particles. Compared to open source PIC and semi-lagrangian codes, there is not an abundant number of open source Eulerian Vlasov in the literature that are not based on an expansion method @Tzoufras:2011 @Touati:2014 @Joglekar:2020. Finally, Vlasov equation is a conservation equation of the number of particle in the phase-space. Therefore, using a finite volume method in order to compute it has the advantage of allowing the use numerical schemes that are numerically particle flux conserving compared to other numerical methods. In addition, Four academic Plasma Physics cases are provided :
[…]
