+++ paper.bib
@@ -184,6 +184,5 @@ issn = {0010-4655},
 doi = {https://doi.org/10.1016/j.cpc.2017.09.024},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465517303314},
 author = {J. Derouillat and A. Beck and F. Pérez and T. Vinci and M. Chiaramello and A. Grassi and M. Flé and G. Bouchard and I. Plotnikov and N. Aunai and J. Dargent and C. Riconda and M. Grech},
-keywords = {Plasma kinetic simulation, Particle-In-Cell (PIC), High-performance computing, Laser–plasma interaction, Astrophysical plasmas},
-Program title: Smilei (version 3.2) Program Files doi: http://dx.doi.org/10.17632/gsn4x6mbrg.1 Licensing provisions: This version of the code is distributed under the GNU General Public License v3 Programming language: C++11, Python 2.7 Nature of the problem: The kinetic simulation of plasmas is at the center of various physics studies, from laser–plasma interaction to astrophysics. To address today’s challenges, a versatile simulation tool requires high-performance computing on massively parallel super-computers. Solution method: The Vlasov–Maxwell system describing the self-consistent evolution of a collisionless plasma is solved using the Particle-In-Cell (PIC) method. Additional physics modules allow to account for additional effects such as collisions and/or ionization. A hybrid MPI-OpenMP strategy, based on a patch-based super-decomposition, allows for efficient cache-use, dynamic load balancing and high-performance on massively parallel super-computers. Additional comments: Repository https://github.com/SmileiPIC/Smilei References: http://www.maisondelasimulation.fr/smilei}
+keywords = {Plasma kinetic simulation, Particle-In-Cell (PIC), High-performance computing, Laser–plasma interaction, Astrophysical plasmas}
 }
