+++ paper.bib
@@ -77,6 +77,13 @@
   author = {Fromm, J. E.}
 }
 
+@book{GaleevSagdeev:1969,
+  author = {Sagdeev, R. Z. and Galeev, A. A.},
+  title = "{Nonlinear Plasma Theory}",
+  year = "1969",
+  publisher = {W. A. Benjamin, Inc., New York}
+}
+
 @article{BeamWarming:1976,
   title = {An implicit finite-difference algorithm for hyperbolic systems in conservation-law form},
   journal = {Journal of Computational Physics},
@@ -103,6 +110,15 @@
   author = {Van Leer, B.}
 }
 
+@book{Landau:1981,
+  Author = {Landau, L. D. and Lifshitz, E. M.},
+  Series = {A Course of Theoretical Physics},
+  Title = {{Physical kinetics}},
+  Volume = {10},
+  Year = {1981},
+  publisher = {Pergamon Press}
+}
+
 @article{Roe:1986,
   author = {Roe, P L},
   title = {Characteristic-Based Schemes for the Euler Equations},
@@ -114,6 +130,19 @@
   doi = {10.1146/annurev.fl.18.010186.002005},
   URL = {https://doi.org/10.1146/annurev.fl.18.010186.002005},
   eprint = {https://doi.org/10.1146/annurev.fl.18.010186.002005}
+}
+
+@article{Berenger:1994,
+  title = {A perfectly matched layer for the absorption of electromagnetic waves},
+  journal = {Journal of Computational Physics},
+  volume = {114},
+  number = {2},
+  pages = {185-200},
+  year = {1994},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1006/jcph.1994.1159},
+  url = {https://www.sciencedirect.com/science/article/pii/S0021999184711594},
+  author = {Berenger, J.-P.}
 }
 
 @article{Tzoufras:2011,
@@ -144,6 +173,19 @@
 	journal = {New Journal of Physics}
 }
 
+@article{Derouillat:2018,
+  title = {Smilei : A collaborative, open-source, multi-purpose particle-in-cell code for plasma simulation},
+  journal = {Computer Physics Communications},
+  volume = {222},
+  pages = {351-373},
+  year = {2018},
+  issn = {0010-4655},
+  doi = {https://doi.org/10.1016/j.cpc.2017.09.024},
+  url = {https://www.sciencedirect.com/science/article/pii/S0010465517303314},
+  author = {J. Derouillat and A. Beck and F. Pérez and T. Vinci and M. Chiaramello and A. Grassi and M. Flé and G. Bouchard and I. Plotnikov and N. Aunai and J. Dargent and C. Riconda and M. Grech},
+  keywords = {Plasma kinetic simulation, Particle-In-Cell (PIC), High-performance computing, Laser–plasma interaction, Astrophysical plasmas}
+}
+
 @article{Joglekar:2020,
   doi = {10.21105/joss.02182},
   url = {https://doi.org/10.21105/joss.02182},
@@ -155,34 +197,4 @@
   author = {Archis S. Joglekar and Matthew C. Levy},
   title = {VlaPy: A Python package for Eulerian Vlasov-Poisson-Fokker-Planck Simulations},
   journal = {Journal of Open Source Software}
-}
-
-
-@book{GaleevSagdeev:1969,
-  author = {Sagdeev, R. Z. and Galeev, A. A.},
-  title = "{Nonlinear Plasma Theory}",
-  year = "1969",
-  publisher = {W. A. Benjamin, Inc., New York}
-}
-
-@book{Landau:1981,
-  Author = {Landau, L. D. and Lifshitz, E. M.},
-  Series = {A Course of Theoretical Physics},
[…]
