+++ paper.md
@@ -14,8 +14,7 @@ tags:
   - minmod
   - superbee
   - Van Leer
-  - MUSCL1
-  - MUSCL2
+  - MUSCL
   - finite volume numerical schemes for phase-space advection
   - Linear Landau damping
   - Non-linear Landau damping
@@ -38,7 +37,7 @@ bibliography: paper.bib
 
 # Summary
 
-ESVM (ElectroStatic Vlasov-Maxwell) is a Vlasov-Maxwell Fortran 90 Eulerian code parallelized using OpenMP allowing for the study of collisionless plasmas. The different finite volume numerical advection schemes @Godunov:1959 implemented in the code in order to discretize the Vlasov equation are the :
+ESVM (ElectroStatic Vlasov-Maxwell) is a single species 1D-1V Vlasov-Maxwell Fortran 90 code parallelized using OpenMP allowing for the study of collisionless plasmas. Many finite volume @Godunov:1959 numerical advection schemes are implemented in the code in order to discretize the Vlasov equation, namely the :
 - donor-cell scheme i.e. the downwind / upwind scheme @Courant:1952 depending on the advection direction in each phase-space cell, 
 - Lax-Wendroff scheme @LaxWendroff:1960, 
 - Fromm scheme @Fromm:1968,
@@ -48,11 +47,11 @@ ESVM (ElectroStatic Vlasov-Maxwell) is a Vlasov-Maxwell Fortran 90 Eulerian code
 - superbee scheme @Roe:1986 and 
 - two Monotonic Upwind-centered Scheme for Conservation Laws (MUSCL) schemes MUSCL1 and MUSCL2. 
 
-Contrary to the linear second order Lax-Wendroff, Fromm and Beam-Warming schemes, the non-linear second order minmod, superbee, Van Leer and MUSCL schemes make use of a Total Variation Diminishing (TVD) flux limiter with the price of becoming a first order scheme in some phase-space cells to limit the numerical oscillations. The donor-cell scheme is a first order scheme. It has therefore the pros of naturally limiting such eventual oscillations but the cons of being highly numerically diffusive too. In ESVM, the discretized Vlasov equation is coupled with the self-consistent Maxwell-Gauss equation for the electrostatic field or equivalently with the Maxwell-Ampere equation with Poisson equation computed at the first time step. While the 1D discretized Poisson equation needs the inversion of a triangular matrix, the second one makes use of the faster second order finite difference Yee scheme @Yee:1966. Python scripts, using the Matplotlib and Numpy packages, are provided to automatically extract and plot the simulation results that are stored in text files. Compilation rules can be easily modified depending on the user compiler preferences using the provided makefile. 
[…]
