+++ paper.md
@@ -59,33 +59,71 @@ Contrary to the linear second order Lax-Wendroff, Fromm and Beam-Warming schemes
 
 # Equations computed by ESVM
 
-Ions are assumed to be fully ionized with an electrical charge $Z e$, $Z$ being the ion atomic number, and immobile with a homogeneous density $n_i$
-The equations computed by the code are the plasma electron 1D-1V Vlasov equation
+Ions are assumed to be immobile with a homogeneous density $n_i$ and fully ionized with an electrical charge $Z e$ where $Z$ is the ion atomic number and $e$ the elementary charge. The plasma electron distribution function $f_e (x,v_x,t)$ is computed by the code according to the plasma electron 1D-1V Vlasov equation
 \begin{equation}
-\label{eq:vlasov1d1v}
-\displaystyle \frac{\partial f_e}{\partial t} (x,v_x,t) + \displaystyle \frac{\partial }{\partial x} \displaystyle \left ( v_x f_e(x,v_x,t) \right ) - \displaystyle \frac{\partial }{\partial v_x} \displaystyle \left ( \displaystyle \frac{e}{m_e} E_x (x,t) f_e (x,v_x,t)\right ) = 0
+  \label{eq:vlasov1d1v}
+  \displaystyle \frac{\partial f_e}{\partial t} (x,v_x,t) + \displaystyle \frac{\partial }{\partial x} \displaystyle \left ( v_x f_e(x,v_x,t) \right ) - \displaystyle \frac{\partial }{\partial v_x} \displaystyle \left ( \displaystyle \frac{e}{m_e} E_x (x,t) f_e (x,v_x,t)\right ) = 0
 \end{equation}
-and the self-consistent Maxwell-Gauss equation for the electrostatic field 
+that is self-consistently coupled with the Maxwell-Gauss equation 
 \begin{equation}
-\label{eq:gauss}
-\displaystyle \frac{\partial E_x}{\partial x} (x,t) = 4 \pi \displaystyle \left ( Z e n_i - e n_e (x,t) \, d v_x \right ) = 4 \pi \displaystyle \left ( Z e n_i - e \displaystyle \int_{-\infty}^\infty f_e (x,v_x,t) \, d v_x \right )
+  \label{eq:gauss}
+  \displaystyle \frac{\partial E_x}{\partial x} (x,t) = 4 \pi \displaystyle \left ( Z e n_i - e n_e (x,t) \, d v_x \right ) = 4 \pi \displaystyle \left ( Z e n_i - e  \right )
 \end{equation}
-or equivalently, the self-consistent Maxwell-Ampere equation
+or equivalently self-consistently coupled with the Maxwell-Ampere equation
 \begin{equation}
-\displaystyle \frac{\partial E_x }{\partial t } (x,t) = - 4 \pi j_e(x,v_x,t)  = 4 \pi e \displaystyle \int_{-\infty}^\infty f_e (x,v_x,t) v_x \, d v_x
+  \label{eq:ampere}
+  \displaystyle \frac{\partial E_x }{\partial t } (x,t) = - 4 \pi j_e(x,t) 
 \end{equation}
-with Maxwell-Gauss equation \autoref{eq:gauss} computed at the simulation start $t=0$.
-Maxwell-Gauss equation is computed by using the electrostatic potential defined according to 
+with Maxwell-Gauss equation \autoref{eq:gauss} computed at the simulation start $t=0$ only. Indeed, by integrating the plasma electron Vlasov equation \autoref{eq:vlasov1d1v}, one gets the hydrodynamic equation of plasma electron number conservation 
 \begin{equation}
-\label{eq:potential}
[…]
