+++ paper.md
@@ -59,7 +59,7 @@ Contrary to the linear second order Lax-Wendroff, Fromm and Beam-Warming schemes
 
 # Equations computed by ESVM
 
-Ions are assumed to be immobile with a homogeneous density $n_i$ and fully ionized with an electrical charge $Z e$ where $Z$ is the ion atomic number and $e$ the elementary charge. The plasma electron distribution function $f_e (x,v_x,t)$ is computed by the code according to the plasma electron 1D-1V Vlasov equation
+Plasma ions are assumed to be immobile with a homogeneous density $n_i$ and fully ionized with an electrical charge $Z e$ where $Z$ is the ion atomic number and $e$ the elementary charge. The plasma electron distribution function $f_e (x,v_x,t)$ is computed by the code according to the plasma electron 1D-1V Vlasov equation
 \begin{equation}
   \label{eq:vlasov1d1v}
   \displaystyle \frac{\partial f_e}{\partial t} (x,v_x,t) + \displaystyle \frac{\partial }{\partial x} \displaystyle \left ( v_x f_e(x,v_x,t) \right ) - \displaystyle \frac{\partial }{\partial v_x} \displaystyle \left ( \displaystyle \frac{e}{m_e} E_x (x,t) f_e (x,v_x,t)\right ) = 0
@@ -67,33 +67,33 @@ Ions are assumed to be immobile with a homogeneous density $n_i$ and fully ioniz
 that is self-consistently coupled with the Maxwell-Gauss equation 
 \begin{equation}
   \label{eq:gauss}
-  \displaystyle \frac{\partial E_x}{\partial x} (x,t) = 4 \pi \displaystyle \left ( Z e n_i - e n_e (x,t) \, d v_x \right ) = 4 \pi \displaystyle \left ( Z e n_i - e  \right )
+  \displaystyle \frac{\partial E_x}{\partial x} (x,t) = 4 \pi e \displaystyle \left ( Z n_i - n_e (x,t) \right )
 \end{equation}
-or equivalently self-consistently coupled with the Maxwell-Ampere equation
+or, equivalently, self-consistently coupled with the Maxwell-Ampere equation
 \begin{equation}
   \label{eq:ampere}
   \displaystyle \frac{\partial E_x }{\partial t } (x,t) = - 4 \pi j_e(x,t) 
 \end{equation}
-with Maxwell-Gauss equation \autoref{eq:gauss} computed at the simulation start $t=0$ only. Indeed, by integrating the plasma electron Vlasov equation \autoref{eq:vlasov1d1v}, one gets the hydrodynamic equation of plasma electron number conservation 
+with Maxwell-Gauss equation \autoref{eq:gauss} computed at the simulation start $t=0$ only. Indeed, by integrating the plasma electron Vlasov equation \autoref{eq:vlasov1d1v} over the whole velocity space $v_x \in \left [ -\infty,\,\infty \right ] $, one gets the hydrodynamic equation of plasma electron number conservation 
 \begin{equation}
   \label{eq:continuity}
   \displaystyle \frac{\partial n_e}{\partial t} (x,t) + \displaystyle \frac{\partial }{\partial x} \displaystyle \left ( n_e v_e(x,t) \right ) = 0,
 \end{equation}
-which, when injected in the time derivative of Maxwell-Gauss equation \autoref{eq:gauss}, provide Maxwell-Ampere equation if Maxwell-Gauss equation \autoref{eq:gauss} is verified at the simulation start. Here, we have noted 
[…]
