+++ paper.bib
@@ -97,7 +97,7 @@
   author = {Beam, R. M. and Warming, R.F }
 }
 
-@article{VanLeer:1977,
+@article{VanLeerIII:1977,
   title = {Towards the ultimate conservative difference scheme III. Upstream-centered finite-difference schemes for ideal compressible flow},
   journal = {Journal of Computational Physics},
   volume = {23},
@@ -108,6 +108,19 @@
   doi = {https://doi.org/10.1016/0021-9991(77)90094-8},
   url = {https://www.sciencedirect.com/science/article/pii/0021999177900948},
   author = {Van Leer, B.}
+}
+
+@article{VanLeerV:1977,
+  title = {Towards the ultimate conservative difference scheme. V. A second-order sequel to Godunov's method},
+  journal = {Journal of Computational Physics},
+  volume = {32},
+  number = {1},
+  pages = {101-136},
+  year = {1979},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1016/0021-9991(79)90145-1},
+  url = {https://www.sciencedirect.com/science/article/pii/0021999179901451},
+  author = {Bram {van Leer}}
 }
 
 @book{Landau:1981,
@@ -143,6 +156,34 @@
   doi = {https://doi.org/10.1006/jcph.1994.1159},
   url = {https://www.sciencedirect.com/science/article/pii/S0021999184711594},
   author = {Berenger, J.-P.}
+}
+
+@article{Crouseille:2004,
+  title = {Numerical approximation of collisional plasmas by high order methods},
+  journal = {Journal of Computational Physics},
+  volume = {201},
+  number = {2},
+  pages = {546-572},
+  year = {2004},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1016/j.jcp.2004.06.007},
+  url = {https://www.sciencedirect.com/science/article/pii/S0021999104002426},
+  author = {Nicolas Crouseilles and Francis Filbet},
+  keywords = {Landau–Fokker–Planck equation, Vlasov–Poisson system, Finite difference methods}
+}
+
+@article{Duclous:2009,
+  title = {High order resolution of the Maxwell–Fokker–Planck–Landau model intended for ICF applications},
+  journal = {Journal of Computational Physics},
+  volume = {228},
+  number = {14},
+  pages = {5072-5100},
+  year = {2009},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1016/j.jcp.2009.04.005},
+  url = {https://www.sciencedirect.com/science/article/pii/S0021999109001879},
+  author = {Roland Duclous and Bruno Dubroca and Francis Filbet and Vladimir Tikhonchuk},
+  keywords = {High order numerical scheme, Fokker–Planck–Landau, Inertial confinement regime, Magnetic field, Electron transport, Energy deposition}
 }
 
 @article{Tzoufras:2011,
