+++ paper.bib
@@ -158,7 +158,7 @@
   author = {Berenger, J.-P.}
 }
 
-@article{Crouseille:2004,
+@article{Crouseilles:2004,
   title = {Numerical approximation of collisional plasmas by high order methods},
   journal = {Journal of Computational Physics},
   volume = {201},
