+++ paper.bib
@@ -158,6 +158,19 @@
   author = {Berenger, J.-P.}
 }
 
+@article{Liu:1994,
+  title = {Weighted Essentially Non-oscillatory Schemes},
+  journal = {Journal of Computational Physics},
+  volume = {115},
+  number = {1},
+  pages = {200-212},
+  year = {1994},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1006/jcph.1994.1187},
+  url = {https://www.sciencedirect.com/science/article/pii/S0021999184711879},
+  author = {Xu-Dong Liu and Stanley Osher and Tony Chan}
+}
+
 @article{Crouseilles:2004,
   title = {Numerical approximation of collisional plasmas by high order methods},
   journal = {Journal of Computational Physics},
