+++ paper.bib
@@ -123,7 +123,7 @@
   author = {Bram {van Leer}}
 }
 
-@book{Landau:1981,
+@book{LandauLifshitz:1981,
   Author = {Landau, L. D. and Lifshitz, E. M.},
   Series = {A Course of Theoretical Physics},
   Title = {{Physical kinetics}},
