+++ paper.bib
@@ -145,6 +145,15 @@
   eprint = {https://doi.org/10.1146/annurev.fl.18.010186.002005}
 }
 
+@article{Decyk:1987,
+  Author = {Viktor K Decyk},
+  Title = {Simulation of Microscopic Processes in Plasma},
+  Date-Added = {2017-09-18 17:46:59 +0000},
+  Date-Modified = {2017-09-18 17:47:37 +0000},
+  Journal = {Proc. 1987 International Conference on Plasma Physics, Kiev, USSR, April 1987, ed. A G Sitenko [World Scientific, Singapore, 1987] vol. II, p. 1075},
+  Year = {1987}
+}
+
 @article{Berenger:1994,
   title = {A perfectly matched layer for the absorption of electromagnetic waves},
   journal = {Journal of Computational Physics},
