+++ paper.bib
@@ -11,6 +11,22 @@
   year = {1952}
 }
 
+@article{BernsteinGreenKruskal:1957,
+  title = {Exact Nonlinear Plasma Oscillations},
+  author = {Bernstein, Ira B. and Greene, John M. and Kruskal, Martin D.},
+  journal = {Phys. Rev.},
+  volume = {108},
+  issue = {3},
+  pages = {546--550},
+  numpages = {0},
+  year = {1957},
+  month = {Nov},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRev.108.546},
+  url = {https://link.aps.org/doi/10.1103/PhysRev.108.546}
+}
+
+
 @article{Godunov:1959,
   Author = {Godunov, S. K.},
   Title = {Eine Differenzenmethode f\"ur die N\"aherungsberechnung unstetiger L\"osungen der hydrodynamischen Gleichungen},
@@ -63,6 +79,22 @@
   Volume = {14},
   Year = {1966},
   Bdsk-Url-1 = {https://doi.org/10.1109/TAP.1966.1138693}}
+
+@article{Gould:1967,
+  title = {Plasma Wave Echo},
+  author = {Gould, R. W. and O'Neil, T. M. and Malmberg, J. H.},
+  journal = {Phys. Rev. Lett.},
+  volume = {19},
+  issue = {5},
+  pages = {219--222},
+  numpages = {0},
+  year = {1967},
+  month = {Jul},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevLett.19.219},
+  url = {https://link.aps.org/doi/10.1103/PhysRevLett.19.219}
+}
+
 
 @article{Fromm:1968,
   title = {A method for reducing dispersion in convective difference schemes},
