+++ paper.bib
@@ -11,6 +11,22 @@
   year = {1952}
 }
 
+@article{BhatnagarGrossKrook:1954,
+  title = {A Model for Collision Processes in Gases. I. Small Amplitude Processes in Charged and Neutral One-Component Systems},
+  author = {Bhatnagar, P. L. and Gross, E. P. and Krook, M.},
+  journal = {Phys. Rev.},
+  volume = {94},
+  issue = {3},
+  pages = {511--525},
+  numpages = {0},
+  year = {1954},
+  month = {May},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRev.94.511},
+  url = {https://link.aps.org/doi/10.1103/PhysRev.94.511}
+}
+
+
 @article{BernsteinGreenKruskal:1957,
   title = {Exact Nonlinear Plasma Oscillations},
   author = {Bernstein, Ira B. and Greene, John M. and Kruskal, Martin D.},
