+++ paper.bib
@@ -202,6 +202,22 @@
   Year = {1987}
 }
 
+@article{BraamsKarney:1987,
+  title = {Differential form of the collision integral for a relativistic plasma},
+  author = {Braams, Bastiaan J. and Karney, Charles F. F.},
+  journal = {Phys. Rev. Lett.},
+  volume = {59},
+  issue = {16},
+  pages = {1817--1820},
+  numpages = {0},
+  year = {1987},
+  month = {Oct},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevLett.59.1817},
+  url = {https://link.aps.org/doi/10.1103/PhysRevLett.59.1817}
+}
+
+
 @article{Berenger:1994,
   title = {A perfectly matched layer for the absorption of electromagnetic waves},
   journal = {Journal of Computational Physics},
