+++ paper.bib
@@ -26,6 +26,16 @@
   url = {https://link.aps.org/doi/10.1103/PhysRev.94.511}
 }
 
+ S. T.  and  , Dokl. Akad. Nauk SSSR 107, 807 (1956) [Sov. Phys. Dokl. 1, 218 (1957)].
+
+@article{BelaievBudker:1957,
+  author = {Belaiev, S. T. and Budker, G. I.},
+  journal = {Dokl. Akad. Nauk SSSR},
+  volume = {107},
+  issue = {807},
+  pages = {[Sov. Phys. Dokl. 1, 218 (1957)]}
+  year = {1957},
+}
 
 @article{BernsteinGreenKruskal:1957,
   title = {Exact Nonlinear Plasma Oscillations},
