+++ paper.bib
@@ -26,15 +26,12 @@
   url = {https://link.aps.org/doi/10.1103/PhysRev.94.511}
 }
 
- S. T.  and  , Dokl. Akad. Nauk SSSR 107, 807 (1956) [Sov. Phys. Dokl. 1, 218 (1957)].
-
 @article{BelaievBudker:1957,
   author = {Belaiev, S. T. and Budker, G. I.},
   journal = {Dokl. Akad. Nauk SSSR},
   volume = {107},
-  issue = {807},
-  pages = {[Sov. Phys. Dokl. 1, 218 (1957)]}
-  year = {1957},
+  issue = {807 [Sov. Phys. Dokl. 1, 218 (1957)]},
+  year = {1956}
 }
 
 @article{BernsteinGreenKruskal:1957,
@@ -51,7 +48,6 @@
   doi = {10.1103/PhysRev.108.546},
   url = {https://link.aps.org/doi/10.1103/PhysRev.108.546}
 }
-
 
 @article{Godunov:1959,
   Author = {Godunov, S. K.},
