+++ paper.bib
@@ -1,3 +1,13 @@
+@article{Courant:1928,
+  author = {Courant, R. and Friedrichs, K. and Lewy, H.},
+  title = {Über die partiellen Differenzengleichungen der mathematischen},
+  journal = {Physik. Math. Ann.}, 
+  volume = {100}, 
+  pages = {32–74},
+  year = {1928}, 
+  doi = {https://doi.org/10.1007/BF01448839}
+}
+
 @article{Courant:1952,
   author = {Courant, R and Isaacson, E and Rees, M},
   title = {On the solution of nonlinear hyperbolic differential equations by finite differences},
