+++ paper.md
@@ -127,45 +127,45 @@ in order to check the energy conservation in the simulation.
 
 # ESVM units
 
-The code units consist in the commonly used electrostatic units : the electron mass $m_e$ for masses, the elementary charge $e$ for electrical charges, the inverse of the Langmuir plasma electron angular frequency $\omega_{p} = \displaystyle \sqrt{ 4 \pi Z n_i e^2 / m_e}$ for times, the Debye electron screening length $\lambda_{\mathrm{Debye}} = v_{T_e} / \omega_{p}$ and the constant ion density $n_i$ for spatial densities. The resulting normalized electrostatic field and electron distribution function consequently reads $\underline{E}_x = e E_x / m_e \omega_{p} v_{T}$ and $\underline{f}_e = f_e v_{T_e} / n_i$, respectively.
+The code units consist in the commonly used electrostatic units : the electron mass $m_e$ for masses, the elementary charge $e$ for electrical charges, the inverse of the Langmuir plasma electron angular frequency $\omega_{p} = \displaystyle \sqrt{ 4 \pi Z n_i e^2 / m_e}$ for times, the Debye electron screening length $\lambda_{\mathrm{Debye}} = v_{T_e} / \omega_{p}$ and the constant ion density $n_i$ for spatial densities. The resulting normalized electrostatic field and electron distribution function consequently reads $\underline{E_x} = e E_x / m_e \omega_{p} v_{T}$ and $\underline{f_e} = f_e v_{T_e} / n_i$, respectively.
 
 # ESVM numerical stability
 
-The spatial grid cells $\Delta x$ must be chosen lower than the Debye length for the simulations of collisionless plasmas where, by definition, the number of electrons in Debye spheres is infinite. $v_{x,\mathrm{min}}$ and $v_{x,\mathrm{max}}$ should always be chosen sufficiently large in such a way that there is no plasma electrons outside the simulation velocity space during the whole simulation. The simulation velocity bin size $\Delta v_x$ should always be chosen lower than the thermal electron velocity and also sufficiently small to capture the desired Physics. The CFL (from the name of its founder R. Courant, K. Friedrichs and H. Lewy @Courant:1928) stability criterium is taken into account inside the code so that the user just needs to specify in the input deck the scalar parameter $\mathrm{cfl} < 1$ such that the normalized simulation time step respects
[…]
