+++ paper.md
@@ -4,10 +4,16 @@ tags:
   - Fortran 90
   - OpenMP
   - Python
-  - Electrostatic Plasma
-  - Poisson versus Maxwell-Ampere solver
-  - Vlasov equation
-  - donor cell
+  - Electrostatic 
+  - Collisionless
+  - Plasma
+  - Poisson 
+  - Maxwell-Gauss
+  - Maxwell-Ampere
+  - Vlasov
+  - Advection
+  - Finite volume
+  - Donor-cell
   - Lax-Wendroff
   - Beam-Warming
   - Fromm 
@@ -15,7 +21,6 @@ tags:
   - superbee
   - Van Leer
   - MUSCL
-  - finite volume numerical schemes for phase-space advection
   - Linear Landau damping
   - Non-linear Landau damping
   - Two-stream instability
@@ -187,12 +192,12 @@ is a sufficient condition for all numerical schemes implemented in ESVM to be st
 # Provided academic cases
 
 Four well-known Plasma Physics academic cases are provided with ESVM :
-1) the emission of an electrostatic wakefield by a Gaussian (in space and velocity-space) electron beam drifting at a mean velocity higher than the plasma electron thermal velocity; cf. \autoref{fig:electrostatic-wakefield}
+1) the emission of an electrostatic wakefield by a Gaussian electron drifting at a velocity greater than the plasma electron thermal velocity; cf. \autoref{fig:electrostatic-wakefield}
 2) the linear Landau damping of an electrostatic wave; cf. \autoref{fig:linear-landau-damping}, 
 3) the non-linear Landau damping of an electrostatic wave; cf. \autoref{fig:non-linear-landau-damping} and 
 3) the two-stream instability; cf. \autoref{fig:two-stream-instability}.
 
-For each Academic case, an example of input deck is provided together with the corresponding simulation result plots that the code typically generates. For 1), 2) and 3), the simulation is initialized assuming a non-drifting collisionless plasma at Maxwell-Boltzmann equilibrium 
+For each academic case, an example of input deck is provided together with the corresponding simulation result plots that the code typically generates. For 1), 2) and 3), the simulation is initialized assuming a non-drifting collisionless plasma at Maxwell-Boltzmann equilibrium 
 \begin{equation}
   \label{MaxwellBoltzmannEquilibrium}
   \displaystyle \left \{
@@ -205,11 +210,11 @@ that is perturbed with a small perturbation
 \begin{equation}
   \delta f_e (x,t=0)= \displaystyle \frac{ A }{ 2 \pi \delta x \delta v } \exp{ \displaystyle \left [ - \displaystyle \frac{ {(x-x_d)}^2 }{ 2 {\delta x}^2 } \right ] } \exp{ \displaystyle \left [ - \displaystyle \frac{ {(v_x-v_d)}^2 }{ 2 {\delta v}^2 } \right ] },
 \end{equation}
-consisting of an electron beam located at $x_d = x_{\mathrm{min}} + ( x_{\mathrm{max}}-x_{\mathrm{min}} )/8$ with a spatial size $\delta x = \lambda_{\mathrm{Debye}} / 4$ drifting at a velocity $v_d$ with a velocity standard deviation $\delta v = v_{T_e} / 40$ at the simulation start for 1), and a small perturbation 
+consisting of a Gaussian electron located at $x_d = x_{\mathrm{min}} + ( x_{\mathrm{max}}-x_{\mathrm{min}} )/8$ with a spatial standard deviation $\delta x = \lambda_{\mathrm{Debye}} / 4$ drifting at a velocity $v_d$ with a velocity standard deviation $\delta v = v_{T_e} / 40$ at the simulation start for 1), and a small perturbation 
 \begin{equation}
[…]
