+++ paper.md
@@ -21,8 +21,7 @@ tags:
   - superbee
   - Van Leer
   - MUSCL
-  - Linear Landau damping
-  - Non-linear Landau damping
+  - Landau damping
   - Two-stream instability
   - Electrostatic wakefield
 authors:
@@ -202,13 +201,13 @@ For each academic case, an example of input deck is provided together with the c
   \label{MaxwellBoltzmannEquilibrium}
   \displaystyle \left \{
   \begin{array}{l}
-      f_e^{(0)} (x,t=0) = \displaystyle \frac{Z n_i}{ \displaystyle \sqrt{ 2 \pi {v_{T_e}}^2 } } \exp{ \displaystyle \left [ - \displaystyle \frac{ v_x^2 }{ 2 {v_{T_e}}^2 } \right ] }
+      f_e^{(0)} (x,v_x,t=0) = \displaystyle \frac{Z n_i}{ \displaystyle \sqrt{ 2 \pi {v_{T_e}}^2 } } \exp{ \displaystyle \left [ - \displaystyle \frac{ v_x^2 }{ 2 {v_{T_e}}^2 } \right ] }
   \cr E_x^{(0)} (x,t=0) = 0
   \end{array} \right .
 \end{equation}
 that is perturbed with a small perturbation 
 \begin{equation}
-  \delta f_e (x,t=0)= \displaystyle \frac{ A }{ 2 \pi \delta x \delta v } \exp{ \displaystyle \left [ - \displaystyle \frac{ {(x-x_d)}^2 }{ 2 {\delta x}^2 } \right ] } \exp{ \displaystyle \left [ - \displaystyle \frac{ {(v_x-v_d)}^2 }{ 2 {\delta v}^2 } \right ] },
+  \delta f_e (x,v_x,t=0)= \displaystyle \frac{ A }{ 2 \pi \delta x \delta v } \exp{ \displaystyle \left [ - \displaystyle \frac{ {(x-x_d)}^2 }{ 2 {\delta x}^2 } \right ] } \exp{ \displaystyle \left [ - \displaystyle \frac{ {(v_x-v_d)}^2 }{ 2 {\delta v}^2 } \right ] },
 \end{equation}
 consisting of a Gaussian electron located at $x_d = x_{\mathrm{min}} + ( x_{\mathrm{max}}-x_{\mathrm{min}} )/8$ with a spatial standard deviation $\delta x = \lambda_{\mathrm{Debye}} / 4$ drifting at a velocity $v_d$ with a velocity standard deviation $\delta v = v_{T_e} / 40$ at the simulation start for 1), and a small perturbation 
 \begin{equation}
@@ -218,24 +217,24 @@ during a short time interval $\delta t = 6 \pi / \omega_0$ after the simulation 
 \begin{equation}
   \displaystyle \left \{
   \begin{array}{ccccc}
-      f_e (x,t) &=& f_e^{(0)} (x,t) &+& \delta f_e (x,t)
+      f_e (x,v_x,t) &=& f_e^{(0)} (x,v_x,t) &+& \delta f_e (x,v_x,t)
   \cr E_x (x,t) &=& E_x^{(0)} (x,t) &+& \delta E_x (x,t)
-  \end{array} \right . \, \mathrm{with}\, \left | \delta f_e (x,t) \right | \ll f_e^{(0)} (x,t)
+  \end{array} \right . \, \mathrm{with}\, \left | \delta f_e (x,v_x,t) \right | \ll f_e^{(0)} (x,v_x,t)
 \end{equation}
 is still respected during the linear stage of the simulation. 
 
[…]
