+++ paper.md
@@ -191,10 +191,10 @@ is a sufficient condition for all numerical schemes implemented in ESVM to be st
 # Provided academic cases
 
 Four well-known Plasma Physics academic cases are provided with ESVM :
-1) the emission of an electrostatic wakefield by a Gaussian electron drifting at a velocity greater than the plasma electron thermal velocity; cf. \autoref{fig:electrostatic-wakefield}
-2) the linear Landau damping of an electrostatic wave; cf. \autoref{fig:linear-landau-damping}, 
-3) the non-linear Landau damping of an electrostatic wave; cf. \autoref{fig:non-linear-landau-damping} and 
-3) the two-stream instability; cf. \autoref{fig:two-stream-instability}.
+1) the emission of an electrostatic wakefield by a Gaussian electron drifting in a collisionless plasma at Maxwell-Boltzmann equilibrium; cf. \autoref{fig:electrostatic-wakefield}
+2) the linear Landau damping of an electrostatic wave in a collisionless plasma at Maxwell-Boltzmann equilibrium; cf. \autoref{fig:linear-landau-damping}, 
+3) the non-linear Landau damping of an electrostatic wave in a collisionless plasma at Maxwell-Boltzmann equilibrium; cf. \autoref{fig:non-linear-landau-damping} and 
+3) the two-stream instability of two counter-propagating homogeneous Gaussian electron beams of exactly opposite drift velocity with same velocity standard deviation; cf. \autoref{fig:two-stream-instability}.
 
 For each academic case, an example of input deck is provided together with the corresponding simulation result plots that the code typically generates. For 1), 2) and 3), the simulation is initialized assuming a non-drifting collisionless plasma at Maxwell-Boltzmann equilibrium 
 \begin{equation}
@@ -209,11 +209,11 @@ that is perturbed with a small perturbation
 \begin{equation}
   \delta f_e (x,v_x,t=0)= \displaystyle \frac{ A }{ 2 \pi \delta x \delta v } \exp{ \displaystyle \left [ - \displaystyle \frac{ {(x-x_d)}^2 }{ 2 {\delta x}^2 } \right ] } \exp{ \displaystyle \left [ - \displaystyle \frac{ {(v_x-v_d)}^2 }{ 2 {\delta v}^2 } \right ] },
 \end{equation}
-consisting of a Gaussian electron located at $x_d = x_{\mathrm{min}} + ( x_{\mathrm{max}}-x_{\mathrm{min}} )/8$ with a spatial standard deviation $\delta x = \lambda_{\mathrm{Debye}} / 4$ drifting at a velocity $v_d$ with a velocity standard deviation $\delta v = v_{T_e} / 40$ at the simulation start for 1), and a small perturbation consisting of a small amplitude electron plasma wave
+consisting of a Gaussian electron located at $x_d = x_{\mathrm{min}} + ( x_{\mathrm{max}}-x_{\mathrm{min}} )/8$ with a standard deviation $\delta x = \lambda_{\mathrm{Debye}} / 4$ drifting at a velocity $v_d$ with a standard deviation $\delta v = v_{T_e} / 40$ at the simulation start for 1), and a small perturbation consisting of a small amplitude electron plasma wave
 \begin{equation}
   \delta E_x (x,t < \delta t) = A \sin{ \displaystyle \left ( \omega_0 t - k x \right ) }
 \end{equation}
[…]
