+++ paper.bib
@@ -316,6 +316,20 @@
 	journal = {New Journal of Physics}
 }
 
+@article{Debuyl:2014,
+  title = {The vmf90 program for the numerical resolution of the Vlasov equation for mean-field systems},
+  journal = {Computer Physics Communications},
+  volume = {185},
+  number = {6},
+  pages = {1822-1827},
+  year = {2014},
+  issn = {0010-4655},
+  doi = {https://doi.org/10.1016/j.cpc.2014.03.004},
+  url = {https://www.sciencedirect.com/science/article/pii/S0010465514000848},
+  author = {Pierre {de Buyl}},
+  keywords = {Vlasov equation, Hamiltonian Mean-Field model, Single wave model}
+}
+
 @article{Derouillat:2018,
   title = {Smilei : A collaborative, open-source, multi-purpose particle-in-cell code for plasma simulation},
   journal = {Computer Physics Communications},
