+++ sources/library.f90
@@ -110,7 +110,7 @@ subroutine INIT_SIMU(x, vx, f_n)
   ! Two-stream instability test case :
     do l=1,N_vx,1
       do i=1,N_x,1
-        f_n(i,l) = (0.5_PR/sqrt(2._PR*pi*(vs**2._PR)))*&
+        f_n(i,l) = (0.5_PR/sqrt(2._PR*pi*))*&
         & ((1._PR+A*sin(k*x(i)))*exp(-((vx(l)-vd)**2._PR)/2._PR) + &
         &  (1._PR-A*sin(k*x(i)))*exp(-((vx(l)+vd)**2._PR)/2._PR))
        end do
