### 2021-10-05 f01996c
delete vs input parameter
