+++ paper.md
@@ -305,7 +305,64 @@ and
   \label{eq:Eq2}
    \widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} = \displaystyle \frac{4 \pi e}{\iota k_p} \displaystyle \int_{-\infty}^\infty \widehat{\widehat{ \delta \text{f}}}_{e,p}^{(+)}  \displaystyle \left ( \omega,\,v_x \right ) d v_x.
 \end{equation}
-Injecting \autoref{Eq1} in \autoref{Eq2}, we obtain the Fourier components of the electrostatic fields
+Injecting \autoref{eq:Eq1} in \autoref{eq:Eq2}, we obtain the Fourier components of the electrostatic fields
+\begin{equation}
+  \widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} \displaystyle \left ( \omega \right ) &=& \displaystyle \frac{4\pi e}{ {k_p}^2 \epsilon \displaystyle \left ( \omega,\,k_p\right )} \displaystyle \int_{-\infty}^\infty \displaystyle \frac{  \widehat{\delta \text{f}}_{e,p} \displaystyle \left ( v_x,\,t=0\right ) }{v_x - \omega / k_p } d v_x
+\end{equation}
+where the plasma electrical permittivity reads
+\begin{equation}
+  \label{eq:plasma_electrical_permittivity}
+  \begin{array}{lll}
+  \epsilon \displaystyle \left ( \omega,\,k \right )   &=&  1 - \displaystyle \frac{4 \pi e^2}{m_e k^2} \displaystyle \int_{-\infty}^\infty \displaystyle \frac{ 1 }{ v_x - \omega / k } \displaystyle \frac{ d f_e^{(0)} }{ d v_x } d v_x
+  \cr                                                                      &=&  1 + \displaystyle \frac{1}{ {\displaystyle \left ( k \lambda_\text{D} \right )}^2} \displaystyle \left \{ 1 + \displaystyle \frac{1}{2} \displaystyle \left [ F \displaystyle \left ( \displaystyle \frac{ \omega / k - v_\text{d} }{ v_\text{th} \sqrt{2} } \right ) +  F \displaystyle \left ( \displaystyle \frac{ \omega / k + v_\text{d} }{ v_\text{th} \sqrt{2} } \right ) \right ] \right \}
+  \end{array}
+\end{equation}
+depending on the plasma dispersion function 
+\begin{equation}
+  \label{eq:plasma_dispersion_function}
+  F \displaystyle \left ( \zeta \right ) = \zeta \mathcal{Z} \displaystyle \left ( \zeta \right ) \text{ and } \mathcal{Z} \displaystyle \left ( \zeta \right ) = \displaystyle \frac{ 1 }{ \sqrt{\pi} } \displaystyle \int_{-\infty}^\infty \displaystyle \frac{ \exp{ \displaystyle \left ( - z^2 \right ) } }{ z - \zeta } d z.
+\end{equation}
+
+Since we are interested in the particular case where $v_d \gg v_{T_e}$, we always have the condition 
+\begin{equation}
+  \label{eq:assumption}
+  \displaystyle \left | \displaystyle \frac{\omega}{k} \pm v_\text{d} \right | \gg v_\text{th} \sqrt{2}
+\end{equation}
+that is fulfilled for any given spatial frequency mode $k$. We thus may use the asymptotic limit
+\begin{equation}
[…]
