+++ paper.md
@@ -312,15 +312,52 @@ Injecting \autoref{eq:Eq1} in \autoref{eq:Eq2}, we obtain the Fourier components
 where the plasma electrical permittivity reads
 \begin{equation}
   \label{eq:plasma_electrical_permittivity}
-  \begin{array}{lll}
-  \epsilon \displaystyle \left ( \omega,\,k \right )   &=&  1 - \displaystyle \frac{4 \pi e^2}{m_e k^2} \displaystyle \int_{-\infty}^\infty \displaystyle \frac{ 1 }{ v_x - \omega / k } \displaystyle \frac{ d f_e^{(0)} }{ d v_x } d v_x
-  \cr                                                                      &=&  1 + \displaystyle \frac{1}{ {\displaystyle \left ( k \lambda_\text{D} \right )}^2} \displaystyle \left \{ 1 + \displaystyle \frac{1}{2} \displaystyle \left [ F \displaystyle \left ( \displaystyle \frac{ \omega / k - v_\text{d} }{ v_\text{th} \sqrt{2} } \right ) +  F \displaystyle \left ( \displaystyle \frac{ \omega / k + v_\text{d} }{ v_\text{th} \sqrt{2} } \right ) \right ] \right \}
-  \end{array}
+  \epsilon \displaystyle \left ( \omega,\,k \right )   =  1 - \displaystyle \frac{4 \pi e^2}{m_e k^2} \displaystyle \int_{-\infty}^\infty \displaystyle \frac{ 1 }{ v_x - \omega / k } \displaystyle \frac{ d f_e^{(0)} }{ d v_x } d v_x =  1 + \displaystyle \frac{1}{ {\displaystyle \left ( k \lambda_\text{D} \right )}^2} \displaystyle \left \{ 1 + \displaystyle \frac{1}{2} \displaystyle \left [ F \displaystyle \left ( \displaystyle \frac{ \omega / k - v_\text{d} }{ v_\text{th} \sqrt{2} } \right ) +  F \displaystyle \left ( \displaystyle \frac{ \omega / k + v_\text{d} }{ v_\text{th} \sqrt{2} } \right ) \right ] \right \}
 \end{equation}
 depending on the plasma dispersion function 
 \begin{equation}
   \label{eq:plasma_dispersion_function}
   F \displaystyle \left ( \zeta \right ) = \zeta \mathcal{Z} \displaystyle \left ( \zeta \right ) \text{ and } \mathcal{Z} \displaystyle \left ( \zeta \right ) = \displaystyle \frac{ 1 }{ \sqrt{\pi} } \displaystyle \int_{-\infty}^\infty \displaystyle \frac{ \exp{ \displaystyle \left ( - z^2 \right ) } }{ z - \zeta } d z.
+\end{equation}
+
+Since we are interested in the particular case where $v_d \gg v_{T_e}$, we always have the condition 
+\begin{equation}
+  \label{eq:assumption}
+  \displaystyle \left | \displaystyle \frac{\omega}{k} \pm v_\text{d} \right | \gg v_\text{th} \sqrt{2}
+\end{equation}
+that is fulfilled for any given spatial frequency mode $k$. We thus may use the asymptotic limit
+\begin{equation}
+  F \displaystyle \left ( \zeta \right ) \underset{|\zeta| \gg 1 }{=}  \iota  \zeta \sqrt{\pi}  \exp{ \displaystyle \left ( - \zeta^2 \right ) }  - 1 - \displaystyle \frac{ 1 }{ 2 \zeta^2} - \displaystyle \frac{ 3 }{ 4 \zeta^4} + ...
+\end{equation}
+that leads to the simpler dispersion relation 
+\begin{equation}
+  \label{eq:plasma_electrical_permittivity_limit}
[…]
