+++ paper.md
@@ -320,48 +320,6 @@ depending on the plasma dispersion function
   F \displaystyle \left ( \zeta \right ) = \zeta \mathcal{Z} \displaystyle \left ( \zeta \right ) \text{ and } \mathcal{Z} \displaystyle \left ( \zeta \right ) = \displaystyle \frac{ 1 }{ \sqrt{\pi} } \displaystyle \int_{-\infty}^\infty \displaystyle \frac{ \exp{ \displaystyle \left ( - z^2 \right ) } }{ z - \zeta } d z.
 \end{equation}
 
-Since we are interested in the particular case where $v_d \gg v_{T_e}$, we always have the condition 
-\begin{equation}
-  \label{eq:assumption}
-  \displaystyle \left | \displaystyle \frac{\omega}{k} \pm v_\text{d} \right | \gg v_\text{th} \sqrt{2}
-\end{equation}
-that is fulfilled for any given spatial frequency mode $k$. We thus may use the asymptotic limit
-\begin{equation}
-  F \displaystyle \left ( \zeta \right ) \underset{|\zeta| \gg 1 }{=}  \iota  \zeta \sqrt{\pi}  \exp{ \displaystyle \left ( - \zeta^2 \right ) }  - 1 - \displaystyle \frac{ 1 }{ 2 \zeta^2} - \displaystyle \frac{ 3 }{ 4 \zeta^4} + ...
-\end{equation}
-that leads to the simpler dispersion relation 
-\begin{equation}
-  \label{eq:plasma_electrical_permittivity_limit}
-  \epsilon \displaystyle \left ( \omega,\,k\right ) \underset{v_\text{d} \gg v_\text{th}}{=} 0 ≤Leftrightarrow 1 - \displaystyle \frac{ {\omega_{p,e}}^2 }{ 2 } \displaystyle \left [  \displaystyle \frac{ 1 }{ {\displaystyle \left ( \omega - k v_\text{d} \right )}^2 } + \displaystyle \frac{ 1 }{ {\displaystyle \left ( \omega + k v_\text{d} \right )}^2 } \right ] = 0
-\end{equation}
-retaining only the main term in the series expansion of the dispersion function \autoref{eq:plasma_dispersion_function} up to the second order. In this limit, dispersion relation thus provides \autoref{eq:plasma_electrical_permittivity_limit} four pure real solutions $\displaystyle \left \{ \omega_1\displaystyle \left ( k \right ),\,\omega_2\displaystyle \left ( k \right ),\,\omega_3\displaystyle \left ( k \right ),\,\omega_4\displaystyle \left ( k \right )  \right \} \in \mathbb{R}^4$ for wavenumber $k$ greater or equal than the critical wavenumber
-\begin{equation}
-  \label{eq:critical_wavenumber}
-  k_c = \displaystyle \frac{\omega_{p,e}}{ v_\text{d} }.
-\end{equation}
-It means that the whole plasma remain stable on space scales smaller than $\lambda_c = 2 \pi / k_c$. However, in the case where $k < k_c$, we find in addition to the two real poles
-\begin{equation}
-  \label{eq:pure_real_poles}
-  \omega_{1/2} \displaystyle \left ( k < k_c \right )= \pm \omega_0 \displaystyle \left ( k \right )
-\end{equation}
-where
-\begin{equation}
-  \label{eq:pure_real_poles}
[…]
