+++ paper.md
@@ -307,7 +307,7 @@ and
 \end{equation}
 Injecting \autoref{eq:Eq1} in \autoref{eq:Eq2}, we obtain the Fourier components of the electrostatic fields
 \begin{equation}
-  \widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} \displaystyle \left ( \omega \right ) &=& \displaystyle \frac{4\pi e}{ {k_p}^2 \epsilon \displaystyle \left ( \omega,\,k_p\right )} \displaystyle \int_{-\infty}^\infty \displaystyle \frac{  \widehat{\delta \text{f}}_{e,p} \displaystyle \left ( v_x,\,t=0\right ) }{v_x - \omega / k_p } d v_x
+  \widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} \displaystyle \left ( \omega \right ) = \displaystyle \frac{4\pi e}{ {k_p}^2 \epsilon \displaystyle \left ( \omega,\,k_p\right )} \displaystyle \int_{-\infty}^\infty \displaystyle \frac{  \widehat{\delta \text{f}}_{e,p} \displaystyle \left ( v_x,\,t=0\right ) }{v_x - \omega / k_p } d v_x
 \end{equation}
 where the plasma electrical permittivity reads
 \begin{equation}
@@ -319,6 +319,55 @@ depending on the plasma dispersion function
   \label{eq:plasma_dispersion_function}
   F \displaystyle \left ( \zeta \right ) = \zeta \mathcal{Z} \displaystyle \left ( \zeta \right ) \text{ and } \mathcal{Z} \displaystyle \left ( \zeta \right ) = \displaystyle \frac{ 1 }{ \sqrt{\pi} } \displaystyle \int_{-\infty}^\infty \displaystyle \frac{ \exp{ \displaystyle \left ( - z^2 \right ) } }{ z - \zeta } d z.
 \end{equation}
+
+Since we are interested in the particular case where $v_d \gg v_{T_e}$, we always have the condition 
+\begin{equation}
+  \label{eq:assumption}
+  \displaystyle \left | \displaystyle \frac{\omega}{k} \pm v_\text{d} \right | \gg v_\text{th} \sqrt{2}
+\end{equation}
+that is fulfilled for any given spatial frequency mode $k$. We thus may use the asymptotic limit
+\begin{equation}
+  F \displaystyle \left ( \zeta \right ) \underset{|\zeta| \gg 1 }{=}  \iota  \zeta \sqrt{\pi}  \exp{ \displaystyle \left ( - \zeta^2 \right ) }  - 1 - \displaystyle \frac{ 1 }{ 2 \zeta^2} - \displaystyle \frac{ 3 }{ 4 \zeta^4} + ...
+\end{equation}
+that leads to the simpler dispersion relation 
+\begin{equation}
+  \label{eq:plasma_electrical_permittivity_limit}
+  \epsilon \displaystyle \left ( \omega,\,k\right ) \underset{v_\text{d} \gg v_\text{th}}{=} 0 ≤Leftrightarrow 1 - \displaystyle \frac{ {\omega_{p,e}}^2 }{ 2 } \displaystyle \left [  \displaystyle \frac{ 1 }{ {\displaystyle \left ( \omega - k v_\text{d} \right )}^2 } + \displaystyle \frac{ 1 }{ {\displaystyle \left ( \omega + k v_\text{d} \right )}^2 } \right ] = 0
+\end{equation}
[…]
