+++ paper.bib
@@ -86,6 +86,13 @@
   year = {1960}
 }
 
+@book{FriedComte:1961,
+  Author = {Fried, B. D. and Conte, S. D.},
+  Series = {The Plasma Dispersion Function},
+  Year = {1961},
+  publisher = {Elsevier}
+}
+
 @article{Dawson:1962,
   author = {Dawson,John },
   title = {One‐Dimensional Plasma Model},
