### 2021-10-06 bdec03e
Update paper.md
