+++ paper.md
@@ -377,12 +377,12 @@ Retaining the main terms in the series expansions of $\mathcal{Z}$ up to the sec
   \label{eq:electrostatic_field_Fourier_modes}
   \widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} \displaystyle \left ( \omega \right ) = \underset{v_d \gg v_{T_e}}{\sim} - \alpha_p A \displaystyle \frac{m_e v_d}{ e } \displaystyle \frac{  {\omega_{p}}^2  }{ \epsilon \displaystyle \left ( \omega,\,k_p\right ) \displaystyle \left ( \omega - k_p v_d \right ) \displaystyle \left ( \omega + k_p v_d \right )  }.
 \end{equation}
-The poles of the Fourier components of the electrostatic fields \autoref{eq:electrostatic_field_Fourier_modes} are thus $\pm k_p v_d $ plus the ones of the plasma electrical permittivity (\autoref{eq:plasma_electrical_permittivity_limit}); Equations \ref{eq:pure_real_poles} and (\autoref{eq:pure_imaginary_poles}). We can know determine the time dependance of the spatial Fourier components of the growing electrostatic field
+The poles of the Fourier components of the electrostatic fields \autoref{eq:electrostatic_field_Fourier_modes} are thus $\pm k_p v_d$ plus the ones of the plasma electrical permittivity (\autoref{eq:plasma_electrical_permittivity_limit}); Equations \ref{eq:pure_real_poles} and (\autoref{eq:pure_imaginary_poles}). We can know determine the time dependance of the spatial Fourier components of the growing electrostatic field
 \begin{equation}
   \label{Inversion_Formula}
   \widehat{\delta \text{E}}_{x,p} \displaystyle \left ( t \right ) = \displaystyle \frac{1}{2 \pi} \displaystyle \int_{\iota R - \infty}^{\iota R + \infty} \widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} \displaystyle \left ( \omega \right )  \exp{\displaystyle \left ( + \iota \omega t \right )} d \omega
 \end{equation} 
-by using the residue theorem with the contour illustrated in \autoref{fig:integration_contour} in order to evaluate the Cauchy principal value of this integral. Since the function $ \widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} \displaystyle \left ( \omega \right )  \exp{\displaystyle \left ( + \iota \omega t \right )} $ is an analytic function of $\omega$ defined in the whole complex plane, we moved  the contour of integration usually taken slightly above the real axis into the lower half-plane sufficiently far beneath the pole $- \iota \delta$ and passing round this pole and round the other poles lying above it in such a way that it doesn't cross any of the poles of the function. We thus obtain
+by using the residue theorem with the contour illustrated in \autoref{fig:integration_contour} in order to evaluate the Cauchy principal value of this integral. Since the function $\widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} \displaystyle \left ( \omega \right )  \exp{\displaystyle \left ( + \iota \omega t \right )}$  is an analytic function of $\omega$ defined in the whole complex plane, we moved the contour of integration usually taken slightly above the real axis into the lower half-plane sufficiently far beneath the pole $- \iota \delta$ and passing round this pole and round the other poles lying above it in such a way that it doesn't cross any of the poles of the function. We thus obtain
 \begin{equation}
   \begin{array}{llllccll}
[…]
