+++ paper.md
@@ -387,7 +387,7 @@ The poles of the Fourier components of the electrostatic fields \autoref{eq:elec
   \label{Inversion_Formula}
   \widehat{\delta \text{E}}_{x,p} \displaystyle \left ( t \right ) = \displaystyle \frac{1}{2 \pi} \displaystyle \int_{\iota R - \infty}^{\iota R + \infty} \widehat{\widehat{\delta \text{E}}}_{x,p}^{(+)} \displaystyle \left ( \omega \right )  \exp{\displaystyle \left ( + \iota \omega t \right )} d \omega
 \end{equation} 
-by using the residue theorem with the contour illustrated in \autoref{fig:integration_contour} in order to evaluate the Cauchy principal value of this integral. Since the function to integrate in \autoref{Inversion_Formula} is an analytic function of $\omega$ defined in the whole complex plane, we moved the contour of integration usually taken slightly above the real axis into the lower half-plane sufficiently far beneath the pole $- \iota \delta$ and passing round this pole and round the other poles lying above it in such a way that it doesn't cross any of the poles of the function. We thus obtain
+by using the residue theorem with the contour illustrated in \autoref{fig:integration_contour} in order to evaluate the Cauchy principal value of this integral : since the function to integrate in \autoref{Inversion_Formula} is an analytic function of $\omega$ defined in the whole complex plane, we moved the contour of integration usually taken slightly above the real axis into the lower half-plane sufficiently far beneath the pole $- \iota \delta$ and passing round this pole and round the other poles lying above it in such a way that it doesn't cross any of the poles of the function. We thus obtain
 \begin{equation}
   \begin{array}{llllccll}
   \widehat{\delta \text{E}}_{x,p} \displaystyle \left ( t \right ) &=&  A & E_0 & \alpha_p &\displaystyle \frac{ \omega_{p} }{ \delta \displaystyle \left ( k_p \right ) }      &\displaystyle \frac{ {\delta \displaystyle \left ( k_p \right )}^2 + {\displaystyle \left ( k_p v_d\right )}^2 }{ {\delta \displaystyle \left ( k_p \right )}^2 + {\omega_0 \displaystyle \left ( k_p \right )}^2 } &  \sinh{ \displaystyle \left [ \delta \displaystyle \left ( k_p \right ) t \right ] }  
