+++ paper.md
@@ -412,7 +412,7 @@ that finally gives according to the Fourier series expansion \autoref{eq:Fourier
 # Perspectives
 
 It is planned in a near future to :
-- provide another Plasma Physics academic simulation about one BGK (from the name of its founder I. B. Bernstein, J. M. Greene and M. D. Kruskal) non linear solution @BernsteinGreenKruskal:1957
+- provide another Plasma Physics academic simulation about one BGK (from the name of its finder I. B. Bernstein, J. M. Greene and M. D. Kruskal) non linear solution @BernsteinGreenKruskal:1957
 - provide a second Plasma Physics academic simulation about Plasma wave echo @Gould:1967
 - implement non-equally spaced phase-space cells
 - implement high order Weighted Essentially Non-Oscillatory (WENO) advection schemes @Liu:1994
@@ -422,7 +422,7 @@ It is planned in a near future to :
 - implement its vectorization
 - store the simulation results in hdf5 files instead of text files
 - implement the Perfectly Matched Layer (PML) technique @Berenger:1994 to absorb the electromagnetic fields at the spatial simulation box boundaries
-- implement a relativistic BGK (from the name of its founder P. L. Bhatnagar, E. P. Gross  and M. Krook) collision operator @BhatnagarGrossKrook:1954
+- implement a relativistic BGK (from the name of its finder P. L. Bhatnagar, E. P. Gross  and M. Krook) collision operator @BhatnagarGrossKrook:1954
 - implement the Belyaev-Budker relativistic collision operator @BelaievBudker:1957 using the Braams-Karney relativistic extension of Rosenbluth potentials @BraamsKarney:1987
 - deploy the code to GPU architectures.
 
