+++ paper.md
@@ -198,8 +198,8 @@ is a sufficient condition for all numerical schemes implemented in ESVM to be st
 
 Four well-known Plasma Physics academic cases are provided with ESVM :
 1) the emission of an electrostatic wakefield by a Gaussian electron; cf. \autoref{fig:electrostatic-wakefield}
-2) the linear Landau damping of an electrostatic wave; cf. \autoref{fig:linear-landau-damping}, 
-3) the non-linear Landau damping of an electrostatic wave; cf. \autoref{fig:non-linear-landau-damping} and 
+2) the linear Landau damping of an electron plasma wave; cf. \autoref{fig:linear-landau-damping}, 
+3) the non-linear Landau damping of an electron plasma wave; cf. \autoref{fig:non-linear-landau-damping} and 
 4) the two-stream instability of two counter-propagating symmetric Gaussian electron beams; cf. \autoref{fig:two-stream-instability}.
 
 For each academic case, an example of input deck is provided together with the corresponding simulation result plots that the code typically generates. For 1), 2) and 3), the simulation is initialized assuming a non-drifting collisionless plasma at Maxwell-Boltzmann equilibrium 
@@ -221,7 +221,7 @@ consisting in a Gaussian electron located at $x_d = x_{\mathrm{min}} + ( x_{\mat
 - with a small perturbation consisting in a small amplitude electron plasma wave
 \begin{equation}
   \label{eq:EPW}
-  \delta E_x (x,t < \delta t) = A \sin{ \displaystyle \left ( \omega_0 t - k x \right ) }
+  \delta E_x (x,t < \delta t) = A \displaystyle \frac{m_e \omega_p v_{T_e}}{e} \sin{ \displaystyle \left ( \omega_0 t - k x \right ) }
 \end{equation}
 propagating during a short time interval $\delta t = 6 \pi / \omega_0$ after the simulation start $t=0$ for 2) and 3). 
 
@@ -261,9 +261,7 @@ on each beam of the form
 \begin{equation}
 \delta f_{e,\pm} \displaystyle \left ( x,v_x,t=0 \right ) = \pm A \sin{\displaystyle \left ( k_1 x \right )  } f_{e,\pm}^{(0)} \displaystyle \left ( x,v_x,t=0 \right )
 \end{equation}
-at the simulation start $t=0$.
-
-Assuming the perturbation \autoref{eq:perturbation} remains small compared to the equilibrium distribution \autoref{eq:EDF} during the simulation, one can linearize the Vlasov equation \autoref{eq:vlasov1d1v} and the self-consistent Maxwell-Gauss equation \autoref{eq:gauss} computed by ESVM. They read
+at the simulation start $t=0$. Assuming the perturbation \autoref{eq:perturbation} remains small compared to the equilibrium distribution \autoref{eq:EDF} during the simulation, one can linearize the Vlasov equation \autoref{eq:vlasov1d1v} and the self-consistent Maxwell-Gauss equation \autoref{eq:gauss} computed by ESVM. They read
 \begin{equation}
   \label{eq:linearized_vlasov1d1v}
   \displaystyle \frac{\partial \delta f_e }{ \partial t} + \displaystyle \frac{\partial }{\partial x} \displaystyle \left ( v_x \delta f_e \right ) - \displaystyle \frac{e}{m_e} \displaystyle \frac{d f_e^{(0)}}{d v_x} \delta E_x= 0
@@ -279,11 +277,7 @@ up to the first order. In order to estimate the linearly growing electrostatic f
[…]
