+++ paper.md
@@ -393,7 +393,7 @@ that finally gives according to the Fourier series expansion \autoref{eq:Fourier
   \end{array}
 \end{equation}
 
-Knowing the electrostatic field \autoref{eq:electrostatic_field}, we may also deduce the perturbed distribution function according to \autoref{eq:linearized_Vlasov1d1v}. It reads 
+Knowing the electrostatic field \autoref{eq:electrostatic_field}, we may also deduce the perturbed distribution function according to \autoref{eq:linearized_vlasov1d1v}. It reads 
 \begin{equation}
   \begin{array}{lll}
   { \delta f}_{e} \displaystyle \left ( x,\, v_x,\,t \right )  &=&  \delta f_{e} \displaystyle \left ( x,\, v_x,\,t=0\right )  + \displaystyle \frac{e}{m_e} \displaystyle \frac{d f_{e}^{(0)}}{d v_x}  \displaystyle \left ( v_x \right ) \displaystyle \int_0^t \delta E_{x} \displaystyle \left [ x + v_x \displaystyle \left ( \tau - t \right ) ,\, \tau \right ] d \tau
@@ -407,7 +407,7 @@ In the limit $k_p v_d \ll \omega_{p}$, they simplify into
 \end{equation}
 and 
 \begin{equation}
-\begin{array}{lcll}
+\begin{array}{llll}
       & & & \displaystyle \frac{e}{m_e} \displaystyle \int_0^t \delta E_{x} \displaystyle \left [ x + v_x \displaystyle \left ( \tau - t \right ) ,\, \tau \right ] d \tau  
  \cr &\underset{k_1 v_d \ll \omega_{p}}{\sim}&  A \displaystyle \frac{ v_d}{ 1 - {\displaystyle \left ( \displaystyle \frac{ k_1 v_x }{ \omega_{p} } \right )}^2  } & \displaystyle \left \{ \displaystyle \frac{k_1 v_x }{ \omega_p } \sin{ \displaystyle \left ( \omega_{p} t \right ) } \;\;\; \cos{ \displaystyle \left ( k_1 x \right ) } - \displaystyle \left [ \cos{ \displaystyle \left ( \omega_{p} t \right ) } \;\,\;- 1 \right ]  \sin{ \displaystyle \left ( k_1 x \right ) } \right \}
  \cr &+& A  \displaystyle \frac{ v_d }{ 1 +  {\displaystyle \left ( \displaystyle \frac{ v_x }{ v_d} \right )}^2 } & \displaystyle \left \{ - \displaystyle \frac{ v_x }{ v_d } \, \sinh{ \displaystyle \left (k_1 v_d t \right ) } \cos{ \displaystyle \left ( k_1 x \right ) } + \displaystyle \left [ \cosh{ \displaystyle \left ( k_1 v_d  t \right ) } - 1 \right ] \sin{ \displaystyle \left ( k_1 x \right ) } \right \}.
@@ -415,17 +415,17 @@ and
 \end{equation}
 We thus deduce in this limit
 \begin{equation}
-\begin{array}{lcl}
-\delta n_e \displaystyle \left ( x,\, t \right ) &=& \displaystyle \int_{-\infty}^\infty \delta f_e \displaystyle\left ( x,\,v_x,\,t \right ) d v_x
[…]
