+++ paper.md
@@ -446,7 +446,7 @@ corresponds to space-charge oscillations of stationary electrostatic plasma wave
   \cr \delta E_{\text{ins}}  \displaystyle \left ( x,\,t \right )       &\underset{k_1 v_d \ll \omega_{p}}{\sim} &  2 A & E_0 &\displaystyle \frac{k_1 v_d}{\omega_{p}} & \sinh{ \displaystyle \left ( k_1 v_d t \right ) } & \sin{ \displaystyle \left ( k_1 x \right ) }
 \end{array} \right .
 \end{equation}
-corresponding to the exponentially growing electrostatic field due to the two-stream instability. These latter growing electron density, current density and electrostatic field perturbations \autoref{eq:my_analytical_solutions} can directly be compared with the ESVM simulation result. We can also check that if $A = 0$ or $v_d = 0$, all quantities cancel. That confirms that, contrary to PIC codes, the two counter-propagating electron beams would continue their propagation without any modification if we do not impose an initial perturbation on which the instability will grow in ESVM. We can now consider the orbit $\displaystyle \left ( x_\ell,\,v_{\ell} \right )$ of one beam electron $\ell \in \displaystyle \left [ 1,\,N_e \right ]$ with an arbitrary initial velocity $v_\ell \displaystyle \left ( t = 0 \right ) = v_0 $ in the beam velocity distribution function and an initial position $x_\ell \displaystyle \left ( t = 0 \right ) = x_0$ close to $x = 0$ such that $ k_1 x_0 \ll 1 $. At the early stage of the instability, the growing electrostatic field component $\delta E_\text{ins}$ is still too small compared to the stationary plasma wave $\delta E_\text{osc}$ oscillating in time at the Langmuir electron angular frequency $\omega_{p}$. On such time scale $\omega_{p} t \sim 1$, the beam electrons are consequently mainly affected by this electrostatic field component
+corresponding to the exponentially growing electrostatic field due to the two-stream instability. These latter growing electron density, current density and electrostatic field perturbations \autoref{eq:my_analytical_solutions} can directly be compared with the ESVM simulation result. We can also check that if $A = 0$ or $v_d = 0$, all quantities cancel. That confirms that, contrary to PIC codes, the two counter-propagating electron beams would continue their propagation without any modification if we do not impose an initial perturbation on which the instability will grow in ESVM. We can now consider the orbit $\displaystyle \left ( x_\ell,\,v_{\ell} \right )$ of one beam electron $\ell \in \displaystyle \left [ 1,\,N_e \right ]$ with an arbitrary initial velocity $v_\ell \displaystyle \left ( t = 0 \right ) = v_0$ in the beam velocity distribution function and an initial position $x_\ell \displaystyle \left ( t = 0 \right ) = x_0$ close to $x = 0$ such that $k_1 x_0 \ll 1$. At the early stage of the instability, the growing electrostatic field component $\delta E_\text{ins}$ is still too small compared to the stationary plasma wave $\delta E_\text{osc}$ oscillating in time at the Langmuir electron angular frequency $\omega_{p}$. On such time scale $\omega_{p} t \sim 1$, the beam electrons are consequently mainly affected by this electrostatic field component
 \begin{equation}
   m_e \displaystyle \frac{d v_\ell }{d t } = - e \delta  E_{\text{osc}} \displaystyle \left ( x_\ell \displaystyle \left ( t \right ),\,t\right )
 \end{equation}
