+++ paper.md
@@ -394,8 +394,7 @@ that finally gives according to the Fourier series expansion \autoref{eq:Fourier
   \cr                                              &+& A & \displaystyle \frac{E_0}{2} & \displaystyle \frac{ {\omega_{p}} }{ \omega_0 \displaystyle \left ( k_1 \right ) } & \displaystyle \frac{ {\omega_0 \displaystyle \left ( k_1 \right )}^2 - {\displaystyle \left ( k_1 v_d\right )}^2 }{  {\delta \displaystyle \left ( k_1 \right )}^2 + {\omega_0 \displaystyle \left ( k_1 \right )}^2 } & \sin{ \displaystyle \left [ \omega_0 \displaystyle \left ( k_1 \right ) t \right ] } \sin{\displaystyle \left ( k_1 x \right )}.
   \end{array}
 \end{equation}
-
-Knowing the electrostatic field \autoref{eq:electrostatic_field}, we may also deduce the perturbed distribution function according to \autoref{eq:linearized_vlasov1d1v}. It reads 
+Knowing the electrostatic field \autoref{eq:electrostatic_field}, one may also deduce the perturbed distribution function according to \autoref{eq:linearized_vlasov1d1v}. It reads 
 \begin{equation}
   \begin{array}{lll}
   { \delta f}_{e} \displaystyle \left ( x,\, v_x,\,t \right )  &=&  \delta f_{e} \displaystyle \left ( x,\, v_x,\,t=0\right )  + \displaystyle \frac{e}{m_e} \displaystyle \frac{d f_{e}^{(0)}}{d v_x}  \displaystyle \left ( v_x \right ) \displaystyle \int_0^t \delta E_{x} \displaystyle \left [ x + v_x \displaystyle \left ( \tau - t \right ) ,\, \tau \right ] d \tau
@@ -459,7 +458,7 @@ assuming that $k_1 x_\ell \displaystyle \left ( t \right ) \ll 1$ remains valid 
   \label{eq:mathieu}
   \displaystyle \frac{d^2 x_\ell }{ d u^2 } + \displaystyle \left [ a - 2 q \cos{\displaystyle \left ( 2 u \right )} \right ] x_\ell \displaystyle \left ( u \right ) = 0 
 \end{equation}
-with $a = 0$ and $q = - A k_1 v_d / \omega_{p} $ by doing the change of variable $u \left ( t\right )= \left ( - \pi / 4 \right ) + \left  ( \omega_{p,e} t / 2 \right ) $, we deduce
+with $a = 0$ and $q = - A k_1 v_d / \omega_{p}$ by doing the change of variable $u \left ( t\right )= \left ( - \pi / 4 \right ) + \left  ( \omega_{p,e} t / 2 \right )$, we deduce
 \begin{equation}
   \label{trajectory_01}
   k_1 x_\ell \displaystyle \left ( t \right ) = k_1 x_c c_{e,0} \displaystyle \left [ q,\,u \left ( t\right ) \right ] +  k_1 x_s s_{e,0} \displaystyle \left [ q,\,u \left ( t\right ) \right ]
@@ -475,8 +474,8 @@ with
[…]
