+++ paper.md
@@ -250,7 +250,8 @@ that is a solution of the Vlasov Equation \autoref{eq:vlasov1d1v} and that doesn
 \end{equation}
 according to Maxwell-Gauss Equation \autoref{eq:gauss}. If one computes the Vlasov-Maxwell set of Equations $\{$\autoref{eq:vlasov1d1v}, \autoref{eq:gauss}$\}$ exactly, initializing it with the two-stream equilibrium distribution function \autoref{eq:EDF} without any perturbation, the counter-propagating electron beams would continue their propagation through the immobile plasma ions without any modification. In order to observe the two-stream instability, 
 \begin{equation}
-f_e \displaystyle \left ( x,v_x,t=0\right ) = f_e^{(0)} \displaystyle \left ( x,v_x,t=0 \right ) + \delta f_{e} \displaystyle \left ( x,v_x,t=0 \right ),
+  \label{eq:perturbed_beams}
+  f_e \displaystyle \left ( x,v_x,t=0\right ) = f_e^{(0)} \displaystyle \left ( x,v_x,t=0 \right ) + \delta f_{e} \displaystyle \left ( x,v_x,t=0 \right ),
 \end{equation}
 is initialized instead by adding a small perturbation 
 \begin{equation}
@@ -261,9 +262,9 @@ on each beam of the form
 \begin{equation}
 \delta f_{e,\pm} \displaystyle \left ( x,v_x,t=0 \right ) = \pm A \sin{\displaystyle \left ( k_1 x \right )  } f_{e,\pm}^{(0)} \displaystyle \left ( x,v_x,t=0 \right )
 \end{equation}
-at the simulation start $t=0$ with $A = 0.1$, $k_1 = 2 \pi / L_x$ where $L_x= x_{\mathrm{max}} - x_{\mathrm{min}}$ is the simulation box size. $A$, $k_1$ (parameter $k$ in the input-deck) and $v_\text{d}$ should be modified by the user. 
+at the simulation start $t=0$ with $A = 0.1$, $k_1 = 2 \pi / L_x$ (parameter $k$ in the input-deck) where $L_x= x_{\mathrm{max}} - x_{\mathrm{min}}$ can be modified by the user in the input-deck. 
 
-In order to estimate the linearly growing electrostatic field $\delta E_x (x,t)$ in this ESVM simulation, one can linearize the Vlasov equation \autoref{eq:vlasov1d1v} and the self-consistent Maxwell-Gauss equation \autoref{eq:gauss} computed by ESVM assuming the perturbation \autoref{eq:perturbation} remains small compared to the equilibrium distribution \autoref{eq:EDF} during the simulation. They read
+In order to get analytical estimates of growing plasma electron density and mean velocity and electrostatic fields in this ESVM simulation, one can linearize the Vlasov equation \autoref{eq:vlasov1d1v} and the self-consistent Maxwell-Gauss equation \autoref{eq:gauss} computed by ESVM assuming the perturbation \autoref{eq:perturbation} remains small compared to the equilibrium distribution \autoref{eq:EDF} during the simulation. They read
 \begin{equation}
   \label{eq:linearized_vlasov1d1v}
   \displaystyle \frac{\partial \delta f_e }{ \partial t} + \displaystyle \frac{\partial }{\partial x} \displaystyle \left ( v_x \delta f_e \right ) - \displaystyle \frac{e}{m_e} \displaystyle \frac{d f_e^{(0)}}{d v_x} \delta E_x= 0
@@ -309,7 +310,7 @@ and
   \label{eq:Eq2}
[…]
