### 2021-10-06 #3618 @michaeltouati
Dear all,
it seems to me I finished today to improve the code and the paper in order to answer all very useful corrections / suggestions / additions of @tclune and @TomGoffrey.
I hope it will be sufficient.
