### 2021-10-06 michaeltouati/ESVM#11 @tclune
Missing community guidelines
Referencing: openjournals/joss-reviews#3618

JOSS wants: "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"  presumably in something like `contrib.md`.   I think if you poke around at other projects you'll find some suitable boiler plate language for this.
