### 2021-10-06 michaeltouati/ESVM#12 @tclune
No references in paper?
Referencing: openjournals/joss-reviews#3618

JOSS wants:  "References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?"
