### 2021-10-06 d7d67cf
Delete bug_report.md
