+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,19 +1,0 @@
-For problems related with plotting, 
-please make sure first that LaTeX, dvipng and ghostscript are each working and on your PATH 
-for the Matplotlib Python package to be able to render tex fonts.
-
-If the bug persists or if it is related to another problem, follow these steps :
-
-1. Go to 'Issues' on the ESVM main repository branch
-2. Click on 'New issue'
-3. Click on 'Bug report'
-4. Describe the bug the more concisely as possible in the title
-5. Describe with the more details as possible the bug and
-6. Click on 'Submit new issue'
-
-For the step 5., please provide the following information :
-OS: [e.g. Ubuntu 20,04]
-Fortran compiler [e.g. gfortran 11.2.0]
-Python version [e.g. Python 3.7.11]
-Matplotlib Python package version [e.g. 3.4.3]
-Numpy Python package version [e.g., 1.21.2]
