+++ readme.md
@@ -1,4 +1,4 @@
-Written by Michael J TOUATI - https://github.com/michaeltouati
+Written by Michael J TOUATI - mtouati@clpu.es
 
 # Compiling the code
 
