### 2021-10-06 michaeltouati/ESVM#12 @michaeltouati
I think it will be better if I replace the paper.md file and the paper.bib file in the repository directly by the .pdf file generated with the JOSS tool once the paper will be published (hoping it will)
