+++ paper.md
@@ -131,11 +131,11 @@ in order to check the energy conservation in the simulation.
 
 # ESVM units
 
-The code units consist in the commonly used electrostatic units : the electron mass $m_e$ for masses, the elementary charge $e$ for electrical charges, the inverse of the Langmuir plasma electron angular frequency $\omega_{p} = \displaystyle \sqrt{ 4 \pi Z n_i e^2 / m_e}$ for times, the Debye electron screening length $\lambda_{\mathrm{Debye}} = v_{T_e} / \omega_{p}$ and the constant electron density $n_0 = Z n_i$ for spatial densities. $v_{T_e}$ is therefore an important unit parameter of normalization since it fixes indirectly the space unit. It can be defined more generally as the plasma electron velocity distribution standard deviation if the plasma is not initialized at Maxwell-Boltzmann thermodynamic equilibrium; cf. \autoref{eq:internal_energy}. Injecting these units in the equations computed by the code, one deduces the resulting normalized electrostatic field and electron distribution function that consequently reads $\underline{E_x} = e E_x / m_e \omega_{p} v_{T_e}$ and $\underline{f_e} = f_e v_{T_e} / n_0$, respectively.
+The code units consist in the commonly used electrostatic units : the electron mass $m_e$ for masses, the elementary charge $e$ for electrical charges, the inverse of the Langmuir plasma electron angular frequency $\omega_{p} = \displaystyle \sqrt{ 4 \pi Z n_i e^2 / m_e}$ for times, the Debye electron screening length $\lambda_{\mathrm{Debye}} = v_{T_{e,0}} / \omega_{p}$ and the constant electron density $n_0 = Z n_i$ for spatial densities. $v_{T_{e,0}}$ is therefore an important unit parameter of normalization since it fixes indirectly the space unit. It can be defined more generally as the plasma electron velocity distribution standard deviation if the plasma is not initialized at Maxwell-Boltzmann thermodynamic equilibrium; cf. \autoref{eq:internal_energy}. Injecting these units in the equations computed by the code, one deduces the resulting normalized electrostatic field and electron distribution function that consequently reads $\underline{E_x} = e E_x / m_e \omega_{p} v_{T_{e,0}}$ and $\underline{f_e} = f_e v_{T_{e,0}} / n_0$, respectively.
 
 # ESVM numerical stability
 
[…]
