+++ contrib.md
@@ -1,48 +1,18 @@
-Written by Michael J TOUATI - mtouati@clpu.es
+# Contributing to ESVM
 
-# Contribute to ESVM 
+First of all, thank you very much for your interest in contributing to ESVM! 
 
-1) download ESVM from https://github.com/michaeltouati/ESVM or type directly on your terminal : 
+Hans Bethe once said : “We need science education to produce scientists, but we need it equally to create literacy in the public. Man has a fundamental urge to comprehend the world about him, and science gives today the only world picture which we can consider as valid. It gives an understanding of the inside of the atom and of the whole universe, or the peculiar properties of the chemical substances and of the manner in which genes duplicate in biology. An educated layman can, of course, not contribute to science, but can enjoy and participate in many scientific discoveries which as constantly made. Such participation was quite common in the 19th century, but has unhappily declined. Literacy in science will enrich a person's life.”
 
-```sh
-git clone https://github.com/michaeltouati/ESVM.git
-```
+But Richard P. Feynmann also added : “Physics is like sex: sure, it may give some practical results, but that's not why we do it.”
 
-2) If it's been a long time you did the step 1., make sure you have the latest ESVM updates by typing : 
+Being fully in agreement with Richard P. feynman while following the advice of Hans Bethe, it seemed to me natural to develop ESVM with Github in order to try to understand the complexity of plasmas, numerical analysis and High Parallel Computing (HPC) and to make it accessible to the largest public as possible at the same time.
 
-```sh
-git pull
-```
+Here can be found how to :
+- Report or fix a bug
+- Propose or submit a new feature
 
-3) Once you've have modified the ESVM source file(s) to fix a bug or propose a new feature, type from your local ESVM directory : 
-
-```sh
-make clean_all 
-```
-
-4) Type then : 
-
-```sh
-git add . 
-```
-
-if you've added new files
-
-5) Type : 
-
-```sh
-git commit -m "MY COMMIT MESSAGE" -m -a 
-```
-
-describing the more clear and concise as possible the goal of the contribution in "MY COMMIT MESSAGE"
-
-6) Ask for the merge of your local ESVM branch to the main branch by typing : 
-
-```sh
-git push
-```
-
-# Report issues or problems with ESVM 
+# Report a bug using Github's [issues](https://github.com/michaeltouati/ESVM/issues)
 
 For problems related with plotting, please make sure first that LaTeX, dvipng and ghostscript are each working and on your PATH for the Matplotlib Python package to be able to render tex fonts.
 If the bug persists or is related to another problem, follow these steps :
@@ -62,10 +32,36 @@ Matplotlib Python package version [e.g. 3.4.3]
 
 Numpy Python package version [e.g., 1.21.2]
 
-# Seek support to add a new feature in ESVM
+# Propose a new feature using Github's [issues](https://github.com/michaeltouati/ESVM/issues)
 
+If you want to propose a new feature for ESVM, follow these steps :
 1) Go to 'Issues' on the ESVM main repository branch
 2) Click on 'New issue'
 4) Describe the new feature request the more clear and concise as possible in the title starting "Feature request :"
 5) Describe with the more details as possible the new feature request and
 6) Click on 'Submit new issue'
+
+# Fix a bug or submit a new feature
+
+ESVM Uses the Fork and pull model [Github Flow](https://guides.github.com/introduction/flow/index.html).
+Pull requests are the best way to propose changes to the codebase (we use [Github Flow](https://guides.github.com/introduction/flow/index.html)). We actively welcome your pull requests:
+
+1. Fork the repo and create your branch from `master`.
+2. If you've added code that should be tested, add tests.
+
+Please, try to keep the code Fortran 90 standard compliant by : 
[…]
