+++ sources/extract/extract_hydro1D.py
@@ -1,0 +1,171 @@
+# -----------------------------------------------
+# SIMULATION RESULTS PLOTTING FOR 1D1V ESVM SIMU  
+# -----------------------------------------------
+# Written by Michael J Touati
+
+import os
+import numpy as np
+import matplotlib
+import matplotlib.pyplot as plt
+from matplotlib.ticker import MaxNLocator
+from matplotlib import mlab, cm
+
+####################
+# Input parameters #
+####################
+
+font = {'family': 'non-serif',
+        'style':  'normal',
+        'color':  'black',
+        'weight': 'normal',
+        'size': 16,
+        }
+
+#############################
+# Create figures/ directory #
+#############################
+
+dir= os.path.dirname("figures/")
+if not os.path.exists(dir):
+    os.mkdir(dir)
+
+#################################
+# functions used in this script #
+#################################
+
+def plot_1D_hydro_quantity_scalar_plot(N_x, file_name, font_properties, y_label, plot_name):
+    x = []
+    p = []
+    X = np.zeros(N_x)
+    P = np.zeros(N_x)
+    file = open(file_name, 'r')
+    counter = 0
+    for line in file:
+        line      = line.strip()
+        array     = line.split()
+        x.append(float(array[1]))
+        p.append(float(array[2]))
+        counter = counter + 1
+        if counter % N_x == 0:
+            N = int(counter / N_x)
+            for i in range(0,N_x):
+                X[i] = x[(N-1)*N_x+i]
+                P[i] = p[(N-1)*N_x+i]
+            time = int(100.*float(array[0]))/100.
+            print('  t (/omega_p) = '+str(time))
+            fig=plt.figure()
+            plt.rc('text', usetex=True)
+            plt.plot(X, P, linewidth=2, color = 'black')
+            plt.title(r'$t =$'+str(time)+r'$\,\omega_p^{-1}$', fontdict=font_properties)
+            plt.xticks(fontsize=16)
+            plt.xlabel(r'$x\,(\lambda_\mathrm{Debye})$', fontdict=font)
+            plt.xlim([np.amin(X),np.amax(X)])
+            plt.ylabel(y_label, fontdict=font)
+            plt.yticks(fontsize=16)
+            ex_min = np.amin(p)
+            ex_max = np.amax(p)
+            if ( ex_min == ex_max ):
+                if (ex_min == 0.) :
+                    ex_min = -1.
+                    ex_max =  1.
+                else :
+                    ex_min = 0.9 * ex_max
+                    ex_max = 1.1 * ex_max
+            plt.ylim([ex_min,ex_max])
+            fig.savefig(name+str(N)+'.png',bbox_inches='tight')
+            plt.close(fig)
+    file.close()
+
+##########################################
+# Find the number of spatial grid points # 
+# (Nvx,Nx) by reading 'results/fe.dat'   #
+##########################################
+
+print('Search for the number of phase-space cells:')
+t0  = []
+v0  = []
+file = open('results/fe.dat', 'r')
+line = file.readline()
+line = line.strip()
+array = line.split()
+t0.append(float(array[0]))
+v0.append(float(array[1]))
+counter = 0
+for line in file:
+    line      = line.strip()
+    array     = line.split()
+    t0.append(float(array[0]))
+    v0.append(float(array[1]))
[…]
