+++ sources/extract/extract.py
@@ -1,0 +1,417 @@
+# -----------------------------------------------
+# SIMULATION RESULTS PLOTTING FOR 1D1V ESVM SIMU  
+# -----------------------------------------------
+# Written by Michael J Touati
+
+import os
+import numpy as np
+import matplotlib
+import matplotlib.pyplot as plt
+from matplotlib.ticker import MaxNLocator
+from matplotlib import mlab, cm
+
+####################
+# Input parameters #
+####################
+
+font = {'family': 'non-serif',
+        'style':  'normal',
+        'color':  'black',
+        'weight': 'normal',
+        'size': 16,
+        }
+
+cmap_Ex  = 'jet'
+cmap_ne  = 'jet'
+cmap_ve  = 'jet'
+cmap_je  = 'jet'
+cmap_vte = 'jet'
+# cmap_fe = 'gist_ncar'
+# cmap_fe = 'nipy_spectral'
+#cmap_fe = 'gist_stern'
+#cmap_fe = 'gnuplot'
+#cmap_fe = 'gnuplot2'
+cmap_fe = 'Blues'
+#cmap_fe = 'jet'
+
+#############################
+# Create figures/ directory #
+#############################
+
+dir= os.path.dirname("figures/")
+if not os.path.exists(dir):
+    os.mkdir(dir)
+
+#################################
+# functions used in this script #
+#################################
+
+def extract_energy_file(file_name):
+    t0 = []
+    u0 = []
+    file = open(file_name,'r')
+    for line in file:
+        line      = line.strip()
+        array     = line.split()
+        t0.append(float(array[0]))
+        u0.append(float(array[1]))
+    file.close()
+    return [t0,u0]
+
+def plot_1D_hydro_quantity_2Dmap(N_x, file_name, font_properties, colormap, plot_title,plot_name):
+    t = []
+    x = []
+    p = []
+    file = open(file_name, 'r')
+    counter = 0
+    for line in file:
+        line      = line.strip()
+        array     = line.split()
+        t.append(float(array[0]))
+        x.append(float(array[1]))
+        p.append(float(array[2]))
+        counter += 1
+        if counter % N_x == 0:
+          N_t = int(counter / N_x)
+    file.close()
+    T = np.zeros((N_x,N_t))
+    X = np.zeros((N_x,N_t))
+    P = np.zeros((N_x,N_t))
+    for i in range(0,N_x):
+        for n in range(0,N_t):
+            T[i][n] = t[n*Nx+i]
+            X[i][n] = x[n*Nx+i]
+            P[i][n] = p[n*Nx+i]
+    cmap = plt.get_cmap(colormap)
+    Maxval = np.amax(p)
+    Minval = np.amin(p)
+    if ( Minval == Maxval ) :
+        if ( Minval == 0. ) :
+            Maxval =  1.
+            Minval = -1.
+        else :
+            Maxval = 1.1*Maxval
+            Minval = 0.9*Maxval
+    norm = cm.colors.Normalize(vmax=Maxval, vmin=Minval)
+    fig=plt.figure()
+    plt.rc('text', usetex=True)
+    plt.pcolormesh(X,T,P,cmap=cmap,norm=norm,shading='gouraud')
+    cbar=plt.colorbar()
+    cbar.ax.tick_params(labelsize=16)
+    plt.title(plot_title, fontdict=font_properties)
+    plt.xticks(fontsize=16)
[…]

[…]
