+++ contrib.md
@@ -16,11 +16,11 @@ For problems related with plotting, please make sure first LaTeX, dvipng and gho
 4) Describe the bug the more clear and concise as possible in the title starting with "Bug :"
 5) Describe with the more details as possible the bug providing the more pieces of information as possible (input-deck, terminal output and/or screeshot) and more importantly, specify your environment :
 
-OS: [e.g. Ubuntu 20.04]
+OS: [e.g., Ubuntu 20.04]
 
-Fortran compiler [e.g. gfortran 11.2.0]
+Fortran compiler [e.g., gfortran 11.2.0]
 
-Python version [e.g. Python 3.7.11]
+Python version [e.g., Python 3.7.11]
 
 Matplotlib Python package version [e.g., Matplotlib 3.4.3]
 
