### 2021-10-08 #3618 @danielskatz
👋 @rouson - we're now about 7 weeks into the process. How is it going?
