### 2021-10-10 #3618 @rouson
@danielskatz thanks for the new invitation.  I just accepted it.   My initial thoughts are below.  I'll submit these as issues on the repository and will offer pull requests in some cases.  Let me know if I should delete the text below once I've added the corresponding issues on pull requests on the submitted repository.

1. In a [comment] above, @michaeltouati mentions having "made the code Fortran 90 standard compliant" and the repository "About" text describes it as Fortran 95.  Newer Fortran standards are very nearly backward compatible, but do declare some older features obsolescent.  Unless there is a specific reason to limit oneself to Fortran 95 (e.g., if it uses features that were deleted in later standards), I recommend simply describing it as "standard-compliant Fortran."  Better yet, I recommend evaluating whether Fortran 2003, 2008, and 2018 offer anything of value.  A lot has happened in the language since 1990-'95.
2.  I highly recommend evaluating the Fortran Package Manager ([`fpm`]) first released in 2020.  At least for purposes of building, it's possible to replace the 415-line `makefile` with a 4-line `fpm` manifest in less than 5 minutes as I just verified by doing so while running a stopwatch.  I'll post a link to a fork that demonstrates this approach.  In conducting my review, I'll take a look at what additional capabilities the `makefile` and whether `fpm` offers the same -- hopefully at greatly reduced complexity in terms of lines of build-system code.   For now, I can say that `fpm` can also run tests and examples without a lot of the repetitive code in the `makefile`, which appears to be poised to grow proportionately with the number of tests and the growth will comprise making repeated copies of `makefile` instructions.
3. The repository takes surprisingly long to clone, probably owing to the large number of `.png` files.  I strongly recommend against storing this many generated binary files in the repository unless for some reason they are required for execution and/or testing.   Binary files make it difficult or impossible for `git` to optimize space by storing diffs.  Changing one character might mean that `git` has to replace the entire file.  Worse still, the old versions of binary files stay in the repository forever.   The only way to remove them is to revise history, which is a discouraged practice.  I'm not sure of the best solution, but there are tools that perform the necessary repo surgery.  See "[Don't ever commit binary files to git]."
4. Explicit deallocation of `allocatable` entities is unnecessary unless either the compiler is broken or one there is specific reason to do so and that reason is backed up with data on memory system performance.  As my book co-author Jim Xia used to day, "Let the compiler do it's job."
5. I usually recommend against non-default array bounds because there are several scenarios in which the bounds are not preserved.  I have some [demonstration code] online that generates a table of some such scenarios.   If a code uses non-default bounds, it might be good to have a  style guide that prescribes behaviors that will avoid pitfalls.  Preferably such a guide would be enforced by continuous-integration tests.

After putting the above comments in issues and pull requests, I'll add more thoughts.

[`fpm`]: https://github.com/fortran-lang/fpm
[comment]: https://github.com/openjournals/joss-reviews/issues/3618#issuecomment-930530903
[demonstration code]: https://github.com/sourceryinstitute/fidbits/tree/master/src/array-bounds-preservation
[Don't ever commit binary files to git]: https://robinwinslow.uk/dont-ever-commit-binary-files-to-git
