### 2021-10-10 #3618 @michaeltouati
Dear Damian @rouson,
Thank you very much for your first initial thoughts :
1. I just recently switch to 95 because I wanted to use the intrinsic cpu_time fortran timer that were not Fortran 90 standard compliant
2. I'll investigate about fpm. It is true that the makefile can become very large if the number of tests further increase. One simpler solution would be to delete some tests or try to make a loop.
3. I have suppressed all the png files from simulation examples except the ones needed to compile the paper.md Should I suppress all the commit history now? What would be the best way to do it according to you? If I do it, the cloning would become almost instantaneous (all files would represents only few MB)
4. I'll follow your recommendation by deleting the deallocation instructions. It's just that I was thinking the opposite... the more pieces of information one gives to the compiler, the best it is.
5. I needed to declare arrays with non-default array bounds to be able to change the simulation parameters in the input-deck without having to recompile the code, the bounds of the arrays depending on the choice of the user written in the input-deck. I'll have a look at your demonstration code tomorrow and I'll add in the contributing.md file a style guide to avoid such problems.
