### 2021-11-08 #3618 @danielskatz
👋 @rouson - any news?  Did you manage to do this

> I still need to move my comments to issues on the repository and create a pull request with suggestions. I hope to have time for this this evening or at the latest would be this weekend.

Please remember to add `openjournals/joss-reviews#3618` in the issues/PRs you open, so that they are visible here, and whether or not they are open or closed is also visible here.
