### 2021-11-22 #3618 @danielskatz
👋 @michaeltouati - I've suggested some changes in https://github.com/michaeltouati/ESVM/pull/15.  In addition, this paper is too long.  See https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain.  Specifically, please remove the sections:
* Equations computed by ESVM
* ESVM units
* ESVM numerical stability
* ESVM Plasma Physics academic case simulations
* Any corresponding figures

This material could perhaps go in one or more supplementary file(s) in the repository (which you could point to in the JOSS paper).

In addition, please check the references to make sure all journals and books are in Title Case, and are not acronyms (e.g., JETP). This can be done by protecting {L}etters or {Words} in the bibtex entry. Finally, just as a note, in English, there is never a space before a ":".
