### 2021-11-23 #3618 @michaeltouati
I have accepted your pull request @danielskatz .

Also, following your suggestion, I removed the mentioned sections in the JOSS paper and I re-wrote them in separate pdf files in the code repo.

Finally, I have checked and corrected the references to make sure they are all in Title Case.
