### 2021-11-23 #3618 @danielskatz
@michaeltouati - I've created one more PR to try to fix two lists: https://github.com/michaeltouati/ESVM/pull/16
In addition, please remove the 'https://doi.org/' prefix from the DOIs listed above.
