### 2021-11-23 #3618 @michaeltouati
@danielskatz - I've accepted your pull request and removed all "https://doi.org/" from the paper.bib file.
