### 2021-11-23 #3618 @danielskatz
@michaeltouati - unfortunately, you've given the JOSS DOI (for the paper) to zenodo. Here, we need to the DOI of the archived software itself. The JOSS DOI will be for the paper. Let zenodo create its own DOI for the software, and give us that.

And I think the version tag is "esvm-v1.0" - is this correct?
