### 2021-11-23 #3618 @michaeltouati
Yes, "esvm-v1.0" is correct @danielskatz

I've also modified the DOI. it is now:
10.5282/zenodo.5720926
