### 2021-11-23 #3618 @danielskatz
@michaeltouati - unfortunately, that DOI does not resolve currently.  I think something got messed up in the process.  I suggest you either contact zenodo help and see if they can fix the problem, or create a new archive where the doi is automatically generated, the contact zenodo help to ask them to remove the old one.
